Split qgsogrprovider.cpp/.h up to seperate into qgsogrproviderutils and qgsorgprovidermetadata
These files were becoming mammoth, with functions from different
classes all mashed together in the .cpp file, making it hard
to navigate the various classes.
