+++ src/core/providers/ogr/qgsogrdataitems.cpp
@@ -32,6 +32,7 @@
 #include "qgsfieldsitem.h"
 #include "qgszipitem.h"
 #include "qgsabstractdatabaseproviderconnection.h"
+#include "qgsogrprovidermetadata.h"
 
 #include <QFileInfo>
 #include <QTextStream>
