+++ src/core/providers/ogr/qgsogrprovider.cpp
@@ -19,37 +19,20 @@ email                : sherman at mrcc.com
 ///@cond PRIVATE
 
 #include "qgscplerrorhandler.h"
-#include "qgsgdalutils.h"
-#include "qgsogrfeatureiterator.h"
 #include "qgslogger.h"
-#include "qgsmessagelog.h"
-#include "qgslocalec.h"
-#include "qgsfeedback.h"
-#include "qgssettings.h"
+#include "qgsvectorfilewriter.h"
 #include "qgsapplication.h"
-#include "qgsauthmanager.h"
-#include "qgscplhttpfetchoverrider.h"
-#include "qgsdataprovider.h"
-#include "qgsfeature.h"
-#include "qgsfields.h"
-#include "qgsgeometry.h"
-#include "qgscoordinatereferencesystem.h"
-#include "qgsvectorlayerexporter.h"
-#include "qgsdataitemprovider.h"
-#include "qgsogrdataitems.h"
-#include "qgsgeopackagedataitems.h"
-#include "qgswkbtypes.h"
-#include "qgsnetworkaccessmanager.h"
+#include "qgssettings.h"
+#include "qgsogrconnpool.h"
 #include "qgsogrtransaction.h"
-#include "qgsgeopackageprojectstorage.h"
-#include "qgsprojectstorageregistry.h"
-#include "qgsprovidermetadata.h"
-#include "qgsogrdbconnection.h"
-#include "qgsgeopackageproviderconnection.h"
-#include "qgis.h"
-#include "qgsembeddedsymbolrenderer.h"
+#include "qgsogrfeatureiterator.h"
+#include "qgsgdalutils.h"
+#include "qgsfeedback.h"
+#include "qgscplhttpfetchoverrider.h"
 #include "qgsmetadatautils.h"
+#include "qgslocalec.h"
 #include "qgssymbol.h"
+#include "qgsembeddedsymbolrenderer.h"
 #include "qgszipitem.h"
 
 #define CPL_SUPRESS_CPLUSPLUS  //#spellok
@@ -68,51 +51,12 @@ email                : sherman at mrcc.com
 #include <limits>
 #include <memory>
 
-#include <QtDebug>
-#include <QFile>
-#include <QDir>
-#include <QFileInfo>
-#include <QMap>
-#include <QMessageBox>
-#include <QString>
 #include <QTextCodec>
-#include <QStorageInfo>
-#include <QRegularExpression>
-
-
-#ifdef Q_OS_WIN
-#include <windows.h>
-#endif
-#ifdef Q_OS_LINUX
-#include <sys/vfs.h>
-#endif
+#include <QFile>
+#include <QFileInfo>
 
 #define TEXT_PROVIDER_KEY QStringLiteral( "ogr" )
 #define TEXT_PROVIDER_DESCRIPTION QStringLiteral( "OGR data provider" )
-
-static OGRwkbGeometryType ogrWkbGeometryTypeFromName( const QString &typeName );
-
-static bool IsLocalFile( const QString &path );
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-Q_GLOBAL_STATIC_WITH_ARGS( QMutex, sGlobalMutex, ( QMutex::Recursive ) )
-#else
-Q_GLOBAL_STATIC( QRecursiveMutex, sGlobalMutex )
-#endif
-
-
-//! Map a dataset name to the number of opened GDAL dataset objects on it (if opened with GDALOpenWrapper, only for GPKG)
-typedef QMap< QString, int > OpenedDsCountMap;
-Q_GLOBAL_STATIC( OpenedDsCountMap, sMapCountOpenedDS )
-
-QMap< QgsOgrProviderUtils::DatasetIdentification,
-      QList<QgsOgrProviderUtils::DatasetWithLayers *> > QgsOgrProviderUtils::sMapSharedDS;
-
-typedef QHash< GDALDatasetH, bool> DsHandleToUpdateModeHash;
-Q_GLOBAL_STATIC( DsHandleToUpdateModeHash, sMapDSHandleToUpdateMode )
-
-typedef QMap< QString, QDateTime > DsNameToLastModifiedDateMap;
-Q_GLOBAL_STATIC( DsNameToLastModifiedDateMap, sMapDSNameToLastModifiedDate )
 
 bool QgsOgrProvider::convertField( QgsField &field, const QTextCodec &encoding )
 {
@@ -268,79 +212,6 @@ void QgsOgrProvider::repack()
     if ( oldcount != mFeaturesCounted )
       emit dataChanged();
   }
-}
-
-
-Qgis::VectorExportResult QgsOgrProviderMetadata::createEmptyLayer( const QString &uri,
-    const QgsFields &fields,
-    QgsWkbTypes::Type wkbType,
-    const QgsCoordinateReferenceSystem &srs,
-    bool overwrite,
-    QMap<int, int> &oldToNewAttrIdxMap,
-    QString &errorMessage,
-    const QMap<QString, QVariant> *options )
-{
-  return QgsOgrProvider::createEmptyLayer(
-           uri, fields, wkbType, srs, overwrite,
-           &oldToNewAttrIdxMap, &errorMessage, options
-         );
-}
-
-static QString AnalyzeURI( QString const &uri,
-                           bool &isSubLayer,
-                           int &layerIndex,
-                           QString &layerName,
-                           QString &subsetString,
-                           OGRwkbGeometryType &ogrGeometryTypeFilter,
-                           QStringList &openOptions )
-{
-  isSubLayer = false;
-  layerIndex = 0;
-  layerName = QString();
-  subsetString = QString();
-  ogrGeometryTypeFilter = wkbUnknown;
-  openOptions.clear();
-
-  QgsDebugMsgLevel( "Data source uri is [" + uri + ']', 2 );
-
-  QVariantMap parts = QgsOgrProviderMetadata().decodeUri( uri );
-
-  if ( parts.contains( QStringLiteral( "layerName" ) ) )
-  {
-    layerName = parts.value( QStringLiteral( "layerName" ) ).toString();
-    isSubLayer = !layerName.isEmpty();
-  }
-
-  if ( parts.contains( QStringLiteral( "layerId" ) ) &&
-       parts.value( QStringLiteral( "layerId" ) ) != QVariant() )
-  {
-    bool ok;
-    layerIndex = parts.value( QStringLiteral( "layerId" ) ).toInt( &ok );
-    if ( ok && layerIndex >= 0 )
-      isSubLayer = true;
-    else
-      layerIndex = -1;
-  }
-
-  if ( parts.contains( QStringLiteral( "subset" ) ) )
-  {
-    subsetString = parts.value( QStringLiteral( "subset" ) ).toString();
-  }
-
-  if ( parts.contains( QStringLiteral( "geometryType" ) ) )
-  {
-    ogrGeometryTypeFilter = ogrWkbGeometryTypeFromName( parts.value( QStringLiteral( "geometryType" ) ).toString() );
-  }
-
-  if ( parts.contains( QStringLiteral( "openOptions" ) ) )
-  {
-    openOptions = parts.value( QStringLiteral( "openOptions" ) ).toStringList();
-  }
-
-  const QString fullPath = parts.value( QStringLiteral( "vsiPrefix" ) ).toString()
-                           + parts.value( QStringLiteral( "path" ) ).toString()
-                           + parts.value( QStringLiteral( "vsiSuffix" ) ).toString();
-  return fullPath;
 }
 
 Qgis::VectorExportResult QgsOgrProvider::createEmptyLayer( const QString &uri,
@@ -510,17 +381,17 @@ QgsOgrProvider::QgsOgrProvider( QString const &uri, const ProviderOptions &optio
 
   QgsDebugMsgLevel( "Data source uri is [" + uri + ']', 2 );
 
-  mFilePath = AnalyzeURI( uri,
-                          mIsSubLayer,
-                          mLayerIndex,
-                          mLayerName,
-                          mSubsetString,
-                          mOgrGeometryTypeFilter,
-                          mOpenOptions );
+  mFilePath = QgsOgrProviderUtils::analyzeURI( uri,
+              mIsSubLayer,
+              mLayerIndex,
+              mLayerName,
+              mSubsetString,
+              mOgrGeometryTypeFilter,
+              mOpenOptions );
 
   if ( mFilePath.contains( QLatin1String( "authcfg" ) ) )
   {
-    QRegularExpression authcfgRe( " authcfg='([^']+)'" );
+    QRegularExpression authcfgRe( QStringLiteral( " authcfg='([^']+)'" ) );
     QRegularExpressionMatch match;
     if ( mFilePath.contains( authcfgRe, &match ) )
     {
@@ -795,27 +666,6 @@ QString QgsOgrProvider::ogrWkbGeometryTypeName( OGRwkbGeometryType type ) const
   return geom;
 }
 
-static OGRwkbGeometryType ogrWkbGeometryTypeFromName( const QString &typeName )
-{
-  if ( typeName == QLatin1String( "Point" ) ) return wkbPoint;
-  else if ( typeName == QLatin1String( "LineString" ) ) return wkbLineString;
-  else if ( typeName == QLatin1String( "Polygon" ) ) return wkbPolygon;
-  else if ( typeName == QLatin1String( "MultiPoint" ) ) return wkbMultiPoint;
-  else if ( typeName == QLatin1String( "MultiLineString" ) ) return wkbMultiLineString;
-  else if ( typeName == QLatin1String( "MultiPolygon" ) ) return wkbMultiPolygon;
-  else if ( typeName == QLatin1String( "GeometryCollection" ) ) return wkbGeometryCollection;
-  else if ( typeName == QLatin1String( "None" ) ) return wkbNone;
-  else if ( typeName == QLatin1String( "Point25D" ) ) return wkbPoint25D;
-  else if ( typeName == QLatin1String( "LineString25D" ) ) return wkbLineString25D;
-  else if ( typeName == QLatin1String( "Polygon25D" ) ) return wkbPolygon25D;
-  else if ( typeName == QLatin1String( "MultiPoint25D" ) ) return wkbMultiPoint25D;
-  else if ( typeName == QLatin1String( "MultiLineString25D" ) ) return wkbMultiLineString25D;
-  else if ( typeName == QLatin1String( "MultiPolygon25D" ) ) return wkbMultiPolygon25D;
-  else if ( typeName == QLatin1String( "GeometryCollection25D" ) ) return wkbGeometryCollection25D;
-  QgsDebugMsg( QStringLiteral( "unknown geometry type: %1" ).arg( typeName ) );
-  return wkbUnknown;
-}
-
 void QgsOgrProvider::addSubLayerDetailsToSubLayerList( int i, QgsOgrLayer *layer, bool withFeatureCount ) const
 {
   QString layerName = QString::fromUtf8( layer->name() );
@@ -847,7 +697,7 @@ void QgsOgrProvider::addSubLayerDetailsToSubLayerList( int i, QgsOgrLayer *layer
   QString longDescription;
   if ( mGDALDriverName == QLatin1String( "OAPIF" ) || mGDALDriverName == QLatin1String( "WFS3" ) )
   {
-    longDescription = layer->GetMetadataItem( "TITLE" );
+    longDescription = layer->GetMetadataItem( QStringLiteral( "TITLE" ) );
   }
 
   QgsDebugMsgLevel( QStringLiteral( "id = %1 name = %2 layerGeomType = %3 longDescription = %4" ).arg( i ).arg( layerName ).arg( layerGeomType ). arg( longDescription ), 2 );
@@ -951,7 +801,7 @@ uint QgsOgrProvider::subLayerCount() const
   uint count = layerCount();
 
   QString errCause;
-  QgsOgrLayerUniquePtr layerStyles = QgsOgrProviderUtils::getLayer( mFilePath, "layer_styles", errCause );
+  QgsOgrLayerUniquePtr layerStyles = QgsOgrProviderUtils::getLayer( mFilePath, QStringLiteral( "layer_styles" ), errCause );
   if ( layerStyles )
   {
     count--;
@@ -1462,7 +1312,7 @@ void QgsOgrProvider::loadMetadata()
                               "WHERE md_standard_uri = %2 and reference_scope = %3" ).arg(
                 QgsSqliteUtils::quotedString( mOgrOrigLayer->name() ),
                 QgsSqliteUtils::quotedString( QStringLiteral( "http://mrcc.com/qgis.dtd" ) ),
-                QgsSqliteUtils::quotedString( QStringLiteral( "table" ) ) );                                                         ;
+                QgsSqliteUtils::quotedString( QStringLiteral( "table" ) ) );
 
         if ( QgsOgrLayerUniquePtr l = mOgrOrigLayer->ExecuteSQL( sql.toLocal8Bit().constData() ) )
         {
@@ -1554,57 +1404,10 @@ void QgsOgrProvider::setRelevantFields( bool fetchGeometry, const QgsAttributeLi
   QgsOgrProviderUtils::setRelevantFields( ogrLayer, mAttributeFields.count(), fetchGeometry, fetchAttributes, mFirstFieldIsFid, mSubsetString );
 }
 
-
-void QgsOgrProviderUtils::setRelevantFields( OGRLayerH ogrLayer, int fieldCount,
-    bool fetchGeometry,
-    const QgsAttributeList &fetchAttributes,
-    bool firstAttrIsFid,
-    const QString &subsetString )
-{
-  if ( OGR_L_TestCapability( ogrLayer, OLCIgnoreFields ) )
-  {
-    QVector<const char *> ignoredFields;
-    OGRFeatureDefnH featDefn = OGR_L_GetLayerDefn( ogrLayer );
-    for ( int i = ( firstAttrIsFid ? 1 : 0 ); i < fieldCount; i++ )
-    {
-      if ( !fetchAttributes.contains( i ) )
-      {
-        // add to ignored fields
-        if ( OGRFieldDefnH field = OGR_FD_GetFieldDefn( featDefn, firstAttrIsFid ? i - 1 : i ) )
-        {
-          const char *fieldName = OGR_Fld_GetNameRef( field );
-          // This is implemented a bit in a hacky way, but in case we are acting on a layer
-          // with a subset filter, do not ignore fields that are found in the
-          // where clause. We do this in a rough way, by looking, in a case
-          // insensitive way, if the current field name is in the subsetString,
-          // so we potentially don't ignore fields we could, in situations like
-          // subsetFilter == "foobar = 2", and there's a "foo" or "bar" field.
-          // Better be safe than sorry.
-          // We could argue that OGR_L_SetIgnoredFields() should be aware of
-          // the fields of the attribute filter, and do not ignore them.
-          if ( subsetString.isEmpty() ||
-               subsetString.indexOf( QString::fromUtf8( fieldName ), 0, Qt::CaseInsensitive ) < 0 )
-          {
-            ignoredFields.append( fieldName );
-          }
-        }
-      }
-    }
-
-    if ( !fetchGeometry )
-      ignoredFields.append( "OGR_GEOMETRY" );
-    ignoredFields.append( "OGR_STYLE" ); // not used by QGIS
-    ignoredFields.append( nullptr );
-
-    OGR_L_SetIgnoredFields( ogrLayer, ignoredFields.data() );
-  }
-}
-
 QgsFeatureIterator QgsOgrProvider::getFeatures( const QgsFeatureRequest &request ) const
 {
   return QgsFeatureIterator( new QgsOgrFeatureIterator( static_cast<QgsOgrFeatureSource *>( featureSource() ), true, request, mTransaction ) );
 }
-
 
 unsigned char *QgsOgrProvider::getGeometryPointer( OGRFeatureH fet )
 {
@@ -2429,7 +2232,7 @@ bool QgsOgrProvider::addAttributes( const QList<QgsField> &attributes )
   // been reset to zero, and QgsVectorLayerEditBuffer::commitChanges() would
   // error out because of this.
   // See https://github.com/qgis/QGIS/issues/26840
-  for ( auto field : oldFields )
+  for ( const QgsField &field : oldFields )
   {
     int idx = mAttributeFields.lookupField( field.name() );
     if ( idx >= 0 )
@@ -3419,971 +3222,6 @@ QString QgsOgrProvider::providerKey()
 QString  QgsOgrProvider::description() const
 {
   return TEXT_PROVIDER_DESCRIPTION;
-} //  QgsOgrProvider::description()
-
-
-/**
- *  Convenience function for readily creating file filters.
- *
- *  Given a long name for a file filter and a regular expression, return
- *  a file filter string suitable for use in a QFileDialog::OpenFiles()
- *  call.  The regular express, glob, will have both all lower and upper
- *  case versions added.
- *  \note Copied from qgisapp.cpp.
- *  \todo XXX This should probably be generalized and moved to a standard utility type thingy.
-*/
-static QString createFileFilter_( QString const &longName, QString const &glob )
-{
-  // return longName + " [OGR] (" + glob.toLower() + ' ' + glob.toUpper() + ");;";
-  return longName + " (" + glob.toLower() + ' ' + glob.toUpper() + ");;";
-} // createFileFilter_
-
-
-QString createFilters( const QString &type )
-{
-  //! Database drivers available
-  static QString sDatabaseDrivers;
-  //! Protocol drivers available
-  static QString sProtocolDrivers;
-  //! File filters
-  static QString sFileFilters;
-  //! Directory drivers
-  static QString sDirectoryDrivers;
-  //! Extensions
-  static QStringList sExtensions;
-  //! Directory extensions
-  static QStringList sDirectoryExtensions;
-  //! Wildcards
-  static QStringList sWildcards;
-
-  // if we've already built the supported vector string, just return what
-  // we've already built
-
-  if ( sFileFilters.isEmpty() || sFileFilters.isNull() )
-  {
-    // register ogr plugins
-    QgsApplication::registerOgrDrivers();
-
-    // first get the GDAL driver manager
-    GDALDriverH driver;          // current driver
-    QString driverName;           // current driver name
-
-    // Grind through all the drivers and their respective metadata.
-    // We'll add a file filter for those drivers that have a file
-    // extension defined for them; the others, welll, even though
-    // theoreticaly we can open those files because there exists a
-    // driver for them, the user will have to use the "All Files" to
-    // open datasets with no explicitly defined file name extension.
-    QgsDebugMsgLevel( QStringLiteral( "Driver count: %1" ).arg( OGRGetDriverCount() ), 3 );
-
-    bool kmlFound = false;
-    bool dwgFound = false;
-    bool dgnFound = false;
-
-    for ( int i = 0; i < OGRGetDriverCount(); ++i )
-    {
-      driver = OGRGetDriver( i );
-
-      Q_CHECK_PTR( driver ); // NOLINT
-
-      if ( !driver )
-      {
-        QgsMessageLog::logMessage( QObject::tr( "Unable to get driver %1" ).arg( i ), QObject::tr( "OGR" ) );
-        continue;
-      }
-
-      driverName = GDALGetDriverShortName( driver );
-
-      if ( driverName.startsWith( QLatin1String( "AVCBin" ) ) )
-      {
-        sDirectoryDrivers += QObject::tr( "Arc/Info Binary Coverage" ) + ",AVCBin;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "AVCE00" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Arc/Info ASCII Coverage" ), QStringLiteral( "*.e00" ) );
-        sExtensions << QStringLiteral( "e00" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "BNA" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Atlas BNA" ), QStringLiteral( "*.bna" ) );
-        sExtensions << QStringLiteral( "bna" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "CSV" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Comma Separated Value" ), QStringLiteral( "*.csv" ) );
-        sExtensions << QStringLiteral( "csv" );
-      }
-      else if ( driverName.startsWith( QObject::tr( "DODS" ) ) )
-      {
-        sProtocolDrivers += QLatin1String( "DODS/OPeNDAP,DODS;" );
-      }
-      else if ( driverName.startsWith( QObject::tr( "CouchDB" ) ) )
-      {
-        sProtocolDrivers += QLatin1String( "CouchDB;" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "FileGDB" ) ) )
-      {
-        sDirectoryDrivers += QObject::tr( "ESRI FileGDB" ) + ",FileGDB;";
-        if ( !sDirectoryExtensions.contains( QStringLiteral( "gdb" ) ) )
-          sDirectoryExtensions << QStringLiteral( "gdb" );
-      }
-#if GDAL_VERSION_NUM >= GDAL_COMPUTE_VERSION(3,1,0)
-      else if ( driverName.startsWith( QLatin1String( "FlatGeobuf" ) ) )
-      {
-        sProtocolDrivers += QLatin1String( "FlatGeobuf;" );
-        sFileFilters += createFileFilter_( QObject::tr( "FlatGeobuf" ), QStringLiteral( "*.fgb" ) );
-        sExtensions << QStringLiteral( "fgb" );
-      }
-#endif
-      else if ( driverName.startsWith( QLatin1String( "PGeo" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "ESRI Personal GeoDatabase" ) + ",PGeo;";
-#ifdef Q_OS_WIN
-        sFileFilters += createFileFilter_( QObject::tr( "ESRI Personal GeoDatabase" ), "*.mdb" );
-        sExtensions << "mdb";
-#endif
-      }
-      else if ( driverName.startsWith( QLatin1String( "SDE" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "ESRI ArcSDE" ) + ",SDE;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "ESRI Shapefile" ) ) )
-      {
-        QString exts = GDALGetMetadataItem( driver, GDAL_DMD_EXTENSIONS, "" );
-        sFileFilters += createFileFilter_( QObject::tr( "ESRI Shapefiles" ), exts.contains( "shz" ) ? QStringLiteral( "*.shp *.shz *.shp.zip" ) : QStringLiteral( "*.shp" ) );
-        sExtensions << QStringLiteral( "shp" ) << QStringLiteral( "dbf" );
-        if ( exts.contains( "shz" ) )
-          sExtensions << QStringLiteral( "shz" ) << QStringLiteral( "shp.zip" );
-      }
-      else if ( driverName.startsWith( QObject::tr( "FMEObjects Gateway" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "FMEObjects Gateway" ), QStringLiteral( "*.fdd" ) );
-        sExtensions << QStringLiteral( "fdd" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "GeoJSONSeq" ) ) )
-      {
-        sProtocolDrivers += QLatin1String( "GeoJSON - Newline Delimited;" );
-        sFileFilters += createFileFilter_( QObject::tr( "GeoJSON Newline Delimited JSON" ), QStringLiteral( "*.geojsonl *.geojsons *.nlgeojson *.json" ) );
-        sExtensions << QStringLiteral( "geojsonl" ) << QStringLiteral( "geojsons" ) << QStringLiteral( "nlgeojson" ) << QStringLiteral( "json" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "GeoJSON" ) ) )
-      {
-        sProtocolDrivers += QLatin1String( "GeoJSON;" );
-        sFileFilters += createFileFilter_( QObject::tr( "GeoJSON" ), QStringLiteral( "*.geojson" ) );
-        sExtensions << QStringLiteral( "geojson" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "GeoRSS" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "GeoRSS" ), QStringLiteral( "*.xml" ) );
-        sExtensions << QStringLiteral( "xml" );
-      }
-      else if ( driverName == QLatin1String( "GML" ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Geography Markup Language [GML]" ), QStringLiteral( "*.gml" ) );
-        sExtensions << QStringLiteral( "gml" );
-      }
-      else if ( driverName == QLatin1String( "GMLAS" ) )
-      {
-        continue;
-      }
-      else if ( driverName.startsWith( QLatin1String( "GMT" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Generic Mapping Tools [GMT]" ), QStringLiteral( "*.gmt" ) );
-        sExtensions << QStringLiteral( "gmt" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "GPX" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "GPS eXchange Format [GPX]" ), QStringLiteral( "*.gpx" ) );
-        sExtensions << QStringLiteral( "gpx" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "GPKG" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "GeoPackage" ), QStringLiteral( "*.gpkg" ) );
-        sExtensions << QStringLiteral( "gpkg" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "GRASS" ) ) )
-      {
-        sDirectoryDrivers += QObject::tr( "Grass Vector" ) + ",GRASS;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "IDB" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "Informix DataBlade" ) + ",IDB;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "Interlis 1" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "INTERLIS 1" ), QStringLiteral( "*.itf *.xml *.ili" ) );
-        sExtensions << QStringLiteral( "itf" ) << QStringLiteral( "xml" ) << QStringLiteral( "ili" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "Interlis 2" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "INTERLIS 2" ), QStringLiteral( "*.xtf *.xml *.ili" ) );
-        sExtensions << QStringLiteral( "xtf" ) << QStringLiteral( "xml" ) << QStringLiteral( "ili" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "Ingres" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "Ingres" ) + ",Ingres;";
-      }
-      else if ( driverName == QLatin1String( "KML" ) || driverName == QLatin1String( "LIBKML" ) )
-      {
-        if ( kmlFound )
-          continue;
-        kmlFound = true;
-        sFileFilters += createFileFilter_( QObject::tr( "Keyhole Markup Language [KML]" ), QStringLiteral( "*.kml *.kmz" ) );
-        sExtensions << QStringLiteral( "kml" ) << QStringLiteral( "kmz" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "MapInfo File" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Mapinfo File" ), QStringLiteral( "*.mif *.tab" ) );
-        sExtensions << QStringLiteral( "mif" ) << QStringLiteral( "tab" );
-      }
-      else if ( driverName == QLatin1String( "DGN" ) || driverName == QLatin1String( "DGNV8" ) )
-      {
-        if ( dgnFound )
-          continue;
-        dgnFound = true;
-        sFileFilters += createFileFilter_( QObject::tr( "Microstation DGN" ), QStringLiteral( "*.dgn" ) );
-        sExtensions << QStringLiteral( "dgn" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "MySQL" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "MySQL" ) + ",MySQL;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "MSSQL" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "MSSQL" ) + ",MSSQL;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "OCI" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "Oracle Spatial" ) + ",OCI;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "ODBC" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "ODBC" ) + ",ODBC;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "OGDI" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "OGDI Vectors" ) + ",OGDI;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "OpenFileGDB" ) ) )
-      {
-        sDirectoryDrivers += QObject::tr( "OpenFileGDB" ) + ",OpenFileGDB;";
-        if ( !sDirectoryExtensions.contains( QStringLiteral( "gdb" ) ) )
-          sDirectoryExtensions << QStringLiteral( "gdb" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "PostgreSQL" ) ) )
-      {
-        sDatabaseDrivers += QObject::tr( "PostgreSQL" ) + ",PostgreSQL;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "S57" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "S-57 Base file" ),
-                                           QStringLiteral( "*.000" ) );
-        sExtensions << QStringLiteral( "000" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "SDTS" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Spatial Data Transfer Standard [SDTS]" ),
-                                           QStringLiteral( "*catd.ddf" ) );
-        sWildcards << QStringLiteral( "*catd.ddf" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "SOSI" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Systematic Organization of Spatial Information [SOSI]" ), QStringLiteral( "*.sos" ) );
-        sExtensions << QStringLiteral( "sos" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "SQLite" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "SQLite/SpatiaLite" ), QStringLiteral( "*.sqlite *.db *.sqlite3 *.db3 *.s3db *.sl3" ) );
-        sExtensions << QStringLiteral( "sqlite" ) << QStringLiteral( "db" ) << QStringLiteral( "sqlite3" ) << QStringLiteral( "db3" ) << QStringLiteral( "s3db" ) << QStringLiteral( "sl3" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "SXF" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Storage and eXchange Format" ), QStringLiteral( "*.sxf" ) );
-        sExtensions << QStringLiteral( "sxf" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "UK .NTF" ) ) )
-      {
-        sDirectoryDrivers += QObject::tr( "UK. NTF2" ) + ",UK. NTF;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "TIGER" ) ) )
-      {
-        sDirectoryDrivers += QObject::tr( "U.S. Census TIGER/Line" ) + ",TIGER;";
-      }
-      else if ( driverName.startsWith( QLatin1String( "OGR_VRT" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "VRT - Virtual Datasource" ),
-                                           QStringLiteral( "*.vrt *.ovf" ) );
-        sExtensions << QStringLiteral( "vrt" ) << QStringLiteral( "ovf" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "XPlane" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "X-Plane/Flightgear" ),
-                                           QStringLiteral( "apt.dat nav.dat fix.dat awy.dat" ) );
-        sWildcards << QStringLiteral( "apt.dat" ) << QStringLiteral( "nav.dat" ) << QStringLiteral( "fix.dat" ) << QStringLiteral( "awy.dat" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "Geoconcept" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Geoconcept" ), QStringLiteral( "*.gxt *.txt" ) );
-        sExtensions << QStringLiteral( "gxt" ) << QStringLiteral( "txt" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "DXF" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "AutoCAD DXF" ), QStringLiteral( "*.dxf" ) );
-        sExtensions << QStringLiteral( "dxf" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "ODS" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Open Document Spreadsheet" ), QStringLiteral( "*.ods" ) );
-        sExtensions << QStringLiteral( "ods" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "XLSX" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "MS Office Open XML spreadsheet" ), QStringLiteral( "*.xlsx" ) );
-        sExtensions << QStringLiteral( "xlsx" );
-      }
-      else if ( driverName.endsWith( QLatin1String( "XLS" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "MS Excel format" ), QStringLiteral( "*.xls" ) );
-        sExtensions << QStringLiteral( "xls" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "EDIGEO" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "EDIGEO" ), QStringLiteral( "*.thf" ) );
-        sExtensions << QStringLiteral( "thf" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "NAS" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "NAS - ALKIS" ), QStringLiteral( "*.xml" ) );
-        sExtensions << QStringLiteral( "xml" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "WAsP" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "WAsP" ), QStringLiteral( "*.map" ) );
-        sExtensions << QStringLiteral( "map" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "PCIDSK" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "PCI Geomatics Database File" ), QStringLiteral( "*.pix" ) );
-        sExtensions << QStringLiteral( "pix" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "GPSTrackMaker" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "GPSTrackMaker" ), QStringLiteral( "*.gtm *.gtz" ) );
-        sExtensions << QStringLiteral( "gtm" ) << QStringLiteral( "gtz" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "VFK" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Czech Cadastral Exchange Data Format" ), QStringLiteral( "*.vfk" ) );
-        sExtensions << QStringLiteral( "vfk" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "OSM" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "OpenStreetMap" ), QStringLiteral( "*.osm *.pbf" ) );
-        sExtensions << QStringLiteral( "osm" ) << QStringLiteral( "pbf" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "SUA" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Special Use Airspace Format" ), QStringLiteral( "*.sua" ) );
-        sExtensions << QStringLiteral( "sua" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "OpenAir" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "OpenAir Special Use Airspace Format" ), QStringLiteral( "*.txt" ) );
-        sExtensions << QStringLiteral( "txt" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "PDS" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Planetary Data Systems TABLE" ), QStringLiteral( "*.xml" ) );
-        sExtensions << QStringLiteral( "xml" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "HTF" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Hydrographic Transfer Format" ), QStringLiteral( "*.htf" ) );
-        sExtensions << QStringLiteral( "htf" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "SVG" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Scalable Vector Graphics" ), QStringLiteral( "*.svg" ) );
-        sExtensions << QStringLiteral( "svg" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "ARCGEN" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Arc/Info Generate" ), QStringLiteral( "*.gen" ) );
-        sExtensions << QStringLiteral( "gen" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "PDF" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "Geospatial PDF" ), QStringLiteral( "*.pdf" ) );
-        sExtensions << QStringLiteral( "pdf" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "SEGY" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "SEG-Y" ), QStringLiteral( "*.sgy *.segy" ) );
-        sExtensions << QStringLiteral( "sgy" ) << QStringLiteral( "segy" );
-      }
-      else if ( driverName.startsWith( QLatin1String( "SEGUKOOA" ) ) )
-      {
-        sFileFilters += createFileFilter_( QObject::tr( "SEG-P1" ), QStringLiteral( "*.seg *.seg1 *.sp1" ) );
-        sFileFilters += createFileFilter_( QObject::tr( "UKOOA P1/90" ), QStringLiteral( "*.uko *.ukooa" ) );
-        sExtensions << QStringLiteral( "seg" ) << QStringLiteral( "seg1" ) << QStringLiteral( "sp1" ) << QStringLiteral( "uko" ) << QStringLiteral( "ukooa" );
-      }
-      else
-      {
-        if ( driverName == QLatin1String( "CAD" ) || driverName == QLatin1String( "DWG" ) )
-        {
-          if ( dwgFound )
-            continue;
-          dwgFound = true;
-        }
-
-        QString myGdalDriverExtensions = GDALGetMetadataItem( driver, GDAL_DMD_EXTENSIONS, "" );
-        QString myGdalDriverLongName = GDALGetMetadataItem( driver, GDAL_DMD_LONGNAME, "" );
-        if ( !( myGdalDriverExtensions.isEmpty() || myGdalDriverLongName.isEmpty() ) )
-        {
-#if QT_VERSION < QT_VERSION_CHECK(5, 15, 0)
-          const QStringList splitExtensions = myGdalDriverExtensions.split( ' ', QString::SkipEmptyParts );
-#else
-          const QStringList splitExtensions = myGdalDriverExtensions.split( ' ', Qt::SkipEmptyParts );
-#endif
-          QString glob;
-
-          for ( const QString &ext : splitExtensions )
-          {
-            sExtensions << ext;
-            if ( !glob.isEmpty() )
-              glob += QLatin1Char( ' ' );
-            glob += "*." + ext;
-          }
-
-          if ( driverName == QLatin1String( "JPEG2000" ) ||
-               driverName.startsWith( QLatin1String( "JP2" ) ) )
-          {
-            // Skip over JPEG2000 drivers, as their vector capabilities are just
-            // a marginal use case
-            continue;
-          }
-
-          sFileFilters += createFileFilter_( myGdalDriverLongName, glob );
-
-        }
-        else
-        {
-          // NOP, we don't know anything about the current driver
-          // with regards to a proper file filter string
-          QgsDebugMsgLevel( QStringLiteral( "Unknown driver %1 for file filters." ).arg( driverName ), 2 );
-        }
-      }
-
-    }                          // each loaded OGR driver
-
-    // sort file filters alphabetically
-    QgsDebugMsgLevel( "myFileFilters: " + sFileFilters, 2 );
-#if QT_VERSION < QT_VERSION_CHECK(5, 15, 0)
-    QStringList filters = sFileFilters.split( QStringLiteral( ";;" ), QString::SkipEmptyParts );
-#else
-    QStringList filters = sFileFilters.split( QStringLiteral( ";;" ), Qt::SkipEmptyParts );
-#endif
-    filters.sort();
-    sFileFilters = filters.join( QLatin1String( ";;" ) ) + ";;";
-    QgsDebugMsgLevel( "myFileFilters: " + sFileFilters, 2 );
-
-    // VSIFileHandler (.zip and .gz files) - second
-    //   see http://trac.osgeo.org/gdal/wiki/UserDocs/ReadInZip
-    // Requires GDAL>=1.6.0 with libz support, let's assume we have it.
-    // This does not work for some file types, see VSIFileHandler doc.
-    QgsSettings settings;
-    if ( settings.value( QStringLiteral( "qgis/scanZipInBrowser2" ), "basic" ).toString() != QLatin1String( "no" ) )
-    {
-      sFileFilters.prepend( createFileFilter_( QObject::tr( "GDAL/OGR VSIFileHandler" ), QStringLiteral( "*.zip *.gz *.tar *.tar.gz *.tgz" ) ) );
-      sExtensions << QStringLiteral( "zip" ) << QStringLiteral( "gz" ) << QStringLiteral( "tar" ) << QStringLiteral( "tar.gz" ) << QStringLiteral( "tgz" );
-    }
-
-    // can't forget the all supported case
-    QStringList exts;
-    for ( const QString &ext : std::as_const( sExtensions ) )
-      exts << QStringLiteral( "*.%1 *.%2" ).arg( ext, ext.toUpper() );
-    sFileFilters.prepend( QObject::tr( "All supported files" ) + QStringLiteral( " (%1);;" ).arg( exts.join( QLatin1Char( ' ' ) ) ) );
-
-    // can't forget the default case - first
-    sFileFilters.prepend( QObject::tr( "All files" ) + " (*);;" );
-
-
-    // cleanup
-    if ( sFileFilters.endsWith( QLatin1String( ";;" ) ) ) sFileFilters.chop( 2 );
-
-    QgsDebugMsgLevel( "myFileFilters: " + sFileFilters, 2 );
-  }
-
-  if ( type == QLatin1String( "file" ) )
-  {
-    return sFileFilters;
-  }
-  if ( type == QLatin1String( "database" ) )
-  {
-    return sDatabaseDrivers;
-  }
-  if ( type == QLatin1String( "protocol" ) )
-  {
-    return sProtocolDrivers;
-  }
-  if ( type == QLatin1String( "directory" ) )
-  {
-    return sDirectoryDrivers;
-  }
-  if ( type == QLatin1String( "extensions" ) )
-  {
-    return sExtensions.join( QLatin1Char( '|' ) );
-  }
-  if ( type == QLatin1String( "directory_extensions" ) )
-  {
-    return sDirectoryExtensions.join( QLatin1Char( '|' ) );
-  }
-  if ( type == QLatin1String( "wildcards" ) )
-  {
-    return sWildcards.join( QLatin1Char( '|' ) );
-  }
-  else
-  {
-    return QString();
-  }
-}
-
-QVariantMap QgsOgrProviderMetadata::decodeUri( const QString &uri ) const
-{
-  QString path = uri;
-  QString layerName;
-  QString subset;
-  QString geometryType;
-  QStringList openOptions;
-  QString databaseName;
-
-  int layerId = -1;
-
-  QString vsiPrefix = qgsVsiPrefix( path );
-  QString vsiSuffix;
-  if ( path.startsWith( vsiPrefix, Qt::CaseInsensitive ) )
-  {
-    path = path.mid( vsiPrefix.count() );
-    if ( vsiPrefix == QLatin1String( "/vsizip/" ) )
-    {
-      const QRegularExpression vsiRegex( QStringLiteral( "(?:\\.zip|\\.tar|\\.gz|\\.tar\\.gz|\\.tgz)([^|]*)" ) );
-      QRegularExpressionMatch match = vsiRegex.match( path );
-      if ( match.hasMatch() )
-      {
-        vsiSuffix = match.captured( 1 );
-        path = path.remove( match.capturedStart( 1 ), match.capturedLength( 1 ) );
-      }
-    }
-  }
-  else
-  {
-    vsiPrefix.clear();
-  }
-
-  if ( path.contains( '|' ) )
-  {
-    const QRegularExpression geometryTypeRegex( QStringLiteral( "\\|geometrytype=([a-zA-Z0-9]*)" ), QRegularExpression::PatternOption::CaseInsensitiveOption );
-    const QRegularExpression layerNameRegex( QStringLiteral( "\\|layername=([^|]*)" ), QRegularExpression::PatternOption::CaseInsensitiveOption );
-    const QRegularExpression layerIdRegex( QStringLiteral( "\\|layerid=([^|]*)" ), QRegularExpression::PatternOption::CaseInsensitiveOption );
-    const QRegularExpression subsetRegex( QStringLiteral( "\\|subset=((?:.*[\r\n]*)*)\\Z" ) );
-    const QRegularExpression openOptionRegex( QStringLiteral( "\\|option:([^|]*)" ) );
-
-
-    // we first try to split off the geometry type component, if that's present. That's a known quantity which
-    // will never be more than a-z characters
-    QRegularExpressionMatch match = geometryTypeRegex.match( path );
-    if ( match.hasMatch() )
-    {
-      geometryType = match.captured( 1 );
-      path = path.remove( match.capturedStart( 0 ), match.capturedLength( 0 ) );
-    }
-
-    // next, we try to find and strip out the layerid/layername component. (This logic is based on the assumption
-    // that a layer name doesn't contain a | character!)
-    // we prefer layer names over layer ids, since they are persistent..
-    match = layerNameRegex.match( path );
-    if ( match.hasMatch() )
-    {
-      layerName = match.captured( 1 );
-      path = path.remove( match.capturedStart( 0 ), match.capturedLength( 0 ) );
-    }
-
-    match = layerIdRegex.match( path );
-    if ( match.hasMatch() )
-    {
-      layerId = match.captured( 1 ).toInt();
-      path = path.remove( match.capturedStart( 0 ), match.capturedLength( 0 ) );
-    }
-
-    while ( true )
-    {
-      match = openOptionRegex.match( path );
-      if ( match.hasMatch() )
-      {
-        openOptions << match.captured( 1 );
-        path = path.remove( match.capturedStart( 0 ), match.capturedLength( 0 ) );
-      }
-      else
-      {
-        break;
-      }
-    }
-
-    // lastly, try to parse out the subset component. This is the biggest unknown, because it's
-    // quite possible that a subset string will contain a | character. If we've already parsed
-    // out all the other known |xxx=yyy tags, then we can safely assume that everything from "|subset=" to the
-    // end of the path is part of the subset filter
-    match = subsetRegex.match( path );
-    if ( match.hasMatch() )
-    {
-      subset = match.captured( 1 );
-      path = path.remove( match.capturedStart( 0 ), match.capturedLength( 0 ) );
-    }
-  }
-
-  // Handles DB connections extracting database name if possible
-  // Example: MySQL:database_name,host=localhost,port=3306 authcfg='f8wwfx8'
-  if ( uri.startsWith( QStringLiteral( "MySQL" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "PostgreSQL" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "MSSQL" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "ODBC" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "PGeo" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "SDE" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "OGDI" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "Ingres" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "IDB" ), Qt::CaseSensitivity::CaseInsensitive ) ||
-       uri.startsWith( QStringLiteral( "OCI" ), Qt::CaseSensitivity::CaseInsensitive ) )
-  {
-    auto parts( path.split( ':' ) );
-    if ( parts.count( ) > 1 )
-    {
-      auto dataParts( parts.at( 1 ).split( ',' ) );
-      if ( dataParts.count() > 0 )
-        databaseName = dataParts.at( 0 );
-    }
-  }
-
-  QVariantMap uriComponents;
-  uriComponents.insert( QStringLiteral( "path" ), path );
-  uriComponents.insert( QStringLiteral( "layerName" ), layerName );
-  uriComponents.insert( QStringLiteral( "layerId" ), layerId > -1 && layerName.isEmpty() ? layerId : QVariant() ) ;
-  if ( !subset.isEmpty() )
-    uriComponents.insert( QStringLiteral( "subset" ), subset );
-  if ( !geometryType.isEmpty() )
-    uriComponents.insert( QStringLiteral( "geometryType" ), geometryType );
-  if ( !databaseName.isEmpty() )
-    uriComponents.insert( QStringLiteral( "databaseName" ), databaseName );
-  if ( !openOptions.isEmpty() )
-    uriComponents.insert( QStringLiteral( "openOptions" ), openOptions );
-  if ( !vsiPrefix.isEmpty() )
-    uriComponents.insert( QStringLiteral( "vsiPrefix" ), vsiPrefix );
-  if ( !vsiSuffix.isEmpty() )
-    uriComponents.insert( QStringLiteral( "vsiSuffix" ), vsiSuffix );
-  return uriComponents;
-}
-
-QString QgsOgrProviderMetadata::encodeUri( const QVariantMap &parts ) const
-{
-  const QString vsiPrefix = parts.value( QStringLiteral( "vsiPrefix" ) ).toString();
-  const QString vsiSuffix = parts.value( QStringLiteral( "vsiSuffix" ) ).toString();
-  const QString path = parts.value( QStringLiteral( "path" ) ).toString();
-  const QString layerName = parts.value( QStringLiteral( "layerName" ) ).toString();
-  const QString layerId = parts.value( QStringLiteral( "layerId" ) ).toString();
-  const QString subset = parts.value( QStringLiteral( "subset" ) ).toString();
-  const QString geometryType = parts.value( QStringLiteral( "geometryType" ) ).toString();
-  const QStringList openOptions = parts.value( QStringLiteral( "openOptions" ) ).toStringList();
-  QString uri = vsiPrefix + path + vsiSuffix
-                + ( !layerName.isEmpty() ? QStringLiteral( "|layername=%1" ).arg( layerName ) : !layerId.isEmpty() ? QStringLiteral( "|layerid=%1" ).arg( layerId ) : QString() )
-                + ( !geometryType.isEmpty() ? QStringLiteral( "|geometrytype=%1" ).arg( geometryType ) : QString() );
-  for ( const QString &openOption : openOptions )
-  {
-    uri += QLatin1String( "|option:" );
-    uri += openOption;
-  }
-  if ( !subset.isEmpty() )
-    uri += QStringLiteral( "|subset=%1" ).arg( subset );
-  return uri;
-}
-
-QString QgsOgrProviderUtils::fileVectorFilters()
-{
-  return createFilters( QStringLiteral( "file" ) );
-}
-
-QString QgsOgrProviderUtils::databaseDrivers()
-{
-  return createFilters( QStringLiteral( "database" ) );
-}
-
-QString QgsOgrProviderUtils::protocolDrivers()
-{
-  return createFilters( QStringLiteral( "protocol" ) );
-}
-
-
-QString QgsOgrProviderUtils::directoryDrivers()
-{
-  return  createFilters( QStringLiteral( "directory" ) );
-}
-
-QStringList QgsOgrProviderUtils::fileExtensions()
-{
-  return  createFilters( QStringLiteral( "extensions" ) ).split( '|' );
-}
-
-QStringList QgsOgrProviderUtils::directoryExtensions()
-{
-  return createFilters( QStringLiteral( "directory_extensions" ) ).split( '|' );
-}
-
-QStringList QgsOgrProviderUtils::wildcards()
-{
-  return  createFilters( QStringLiteral( "wildcards" ) ).split( '|' );
-}
-
-
-/**
- * Class factory to return a pointer to a newly created
- * QgsOgrProvider object
- */
-QgsOgrProvider *QgsOgrProviderMetadata::createProvider( const QString &uri, const QgsDataProvider::ProviderOptions &options,
-    QgsDataProvider::ReadFlags flags )
-{
-  return new QgsOgrProvider( uri, options, flags );
-}
-
-/**
- * Creates an empty data source
- * \param uri location to store the file(s)
- * \param format data format (e.g. "ESRI Shapefile")
- * \param vectortype point/line/polygon or multitypes
- * \param attributes a list of name/type pairs for the initial attributes
- * \return true in case of success
-*/
-bool QgsOgrProviderUtils::createEmptyDataSource( const QString &uri,
-    const QString &format,
-    const QString &encoding,
-    QgsWkbTypes::Type vectortype,
-    const QList< QPair<QString, QString> > &attributes,
-    const QgsCoordinateReferenceSystem &srs,
-    QString &errorMessage )
-{
-  QgsDebugMsgLevel( QStringLiteral( "Creating empty vector layer with format: %1" ).arg( format ), 2 );
-  errorMessage.clear();
-
-  QgsApplication::registerOgrDrivers();
-  OGRSFDriverH driver = OGRGetDriverByName( format.toLatin1() );
-  if ( !driver )
-  {
-    return false;
-  }
-
-  QString driverName = GDALGetDriverShortName( driver );
-
-  if ( driverName == QLatin1String( "ESRI Shapefile" ) )
-  {
-    if ( !uri.endsWith( QLatin1String( ".shp" ), Qt::CaseInsensitive ) &&
-         !uri.endsWith( QLatin1String( ".shz" ), Qt::CaseInsensitive ) &&
-         !uri.endsWith( QLatin1String( ".dbf" ), Qt::CaseInsensitive ) )
-    {
-      errorMessage = QObject::tr( "URI %1 doesn't end with .shp or .dbf" ).arg( uri );
-      QgsDebugMsg( errorMessage );
-      return false;
-    }
-
-    // check for duplicate fieldnames
-    QSet<QString> fieldNames;
-    QList<QPair<QString, QString> >::const_iterator fldIt;
-    for ( fldIt = attributes.begin(); fldIt != attributes.end(); ++fldIt )
-    {
-      QString name = fldIt->first.left( 10 );
-      if ( fieldNames.contains( name ) )
-      {
-        errorMessage = QObject::tr( "Duplicate field (10 significant characters): %1" ).arg( name );
-        QgsMessageLog::logMessage( errorMessage, QObject::tr( "OGR" ) );
-        return false;
-      }
-      fieldNames << name;
-    }
-
-    QgsVectorFileWriter::deleteShapeFile( uri );
-  }
-  else
-  {
-    QFile::remove( uri );
-  }
-
-  gdal::dataset_unique_ptr dataSource;
-  dataSource.reset( OGR_Dr_CreateDataSource( driver, uri.toUtf8().constData(), nullptr ) );
-  if ( !dataSource )
-  {
-    errorMessage = QObject::tr( "Creating the data source %1 failed: %2" ).arg( uri, QString::fromUtf8( CPLGetLastErrorMsg() ) );
-    QgsMessageLog::logMessage( errorMessage, QObject::tr( "OGR" ) );
-    return false;
-  }
-
-  //consider spatial reference system
-  OGRSpatialReferenceH reference = nullptr;
-
-  QgsCoordinateReferenceSystem mySpatialRefSys;
-  if ( srs.isValid() )
-  {
-    mySpatialRefSys = srs;
-  }
-  else
-  {
-    mySpatialRefSys.validate();
-  }
-
-  QString myWkt = mySpatialRefSys.toWkt( QgsCoordinateReferenceSystem::WKT_PREFERRED_GDAL );
-
-  if ( !myWkt.isNull()  &&  myWkt.length() != 0 )
-  {
-    reference = OSRNewSpatialReference( myWkt.toLocal8Bit().data() );
-  }
-
-  // Map the qgis geometry type to the OGR geometry type
-  OGRwkbGeometryType OGRvectortype = wkbUnknown;
-  switch ( vectortype )
-  {
-    case QgsWkbTypes::GeometryCollection:
-    case QgsWkbTypes::GeometryCollectionZ:
-    case QgsWkbTypes::GeometryCollectionM:
-    case QgsWkbTypes::GeometryCollectionZM:
-    case QgsWkbTypes::Unknown:
-    {
-      errorMessage = QObject::tr( "Unknown vector type of %1" ).arg( static_cast< int >( vectortype ) );
-      QgsMessageLog::logMessage( errorMessage, QObject::tr( "OGR" ) );
-      return false;
-    }
-
-    default:
-      OGRvectortype = QgsOgrProviderUtils::ogrTypeFromQgisType( vectortype );
-  }
-
-  char **papszOptions = nullptr;
-  if ( driverName == QLatin1String( "ESRI Shapefile" ) )
-  {
-    papszOptions = CSLSetNameValue( papszOptions, "ENCODING", QgsVectorFileWriter::convertCodecNameForEncodingOption( encoding ).toLocal8Bit().data() );
-    // OGR Shapefile fails to create fields if given encoding is not supported by its side
-    // so disable encoding conversion of OGR Shapefile layer
-    CPLSetConfigOption( "SHAPE_ENCODING", "" );
-  }
-
-  OGRLayerH layer;
-  layer = GDALDatasetCreateLayer( dataSource.get(), QFileInfo( uri ).completeBaseName().toUtf8().constData(), reference, OGRvectortype, papszOptions );
-  CSLDestroy( papszOptions );
-
-  if ( !layer )
-  {
-    errorMessage = QString::fromUtf8( CPLGetLastErrorMsg() );
-    QgsMessageLog::logMessage( errorMessage, QObject::tr( "OGR" ) );
-    return false;
-  }
-
-  //create the attribute fields
-
-  QTextCodec *codec = QTextCodec::codecForName( encoding.toLocal8Bit().data() );
-  if ( !codec )
-  {
-    // fall back to "System" codec
-    codec = QTextCodec::codecForLocale();
-    Q_ASSERT( codec );
-  }
-
-  for ( QList<QPair<QString, QString> >::const_iterator it = attributes.begin(); it != attributes.end(); ++it )
-  {
-    QStringList fields = it->second.split( ';' );
-
-    if ( fields.isEmpty() )
-      continue;
-
-    int width = fields.size() > 1 ? fields[1].toInt() : -1;
-    int precision = fields.size() > 2 ? fields[2].toInt() : -1;
-    if ( precision > 0 )
-      width += 1;
-
-    OGRFieldDefnH field;
-    if ( fields[0] == QLatin1String( "Real" ) )
-    {
-      if ( width < 0 )
-        width = 32;
-      if ( precision < 0 )
-        precision = 3;
-
-      field = OGR_Fld_Create( codec->fromUnicode( it->first ).constData(), OFTReal );
-      OGR_Fld_SetWidth( field, width );
-      OGR_Fld_SetPrecision( field, precision );
-    }
-    else if ( fields[0] == QLatin1String( "Integer" ) )
-    {
-      if ( width < 0 || width > 10 )
-        width = 10;
-
-      field = OGR_Fld_Create( codec->fromUnicode( it->first ).constData(), OFTInteger );
-      // limit to 10.  otherwise OGR sets it to 11 and recognizes as OFTDouble later
-      OGR_Fld_SetWidth( field, width );
-    }
-    else if ( fields[0] == QLatin1String( "String" ) )
-    {
-      if ( width < 0 || width > 255 )
-        width = 255;
-
-      field = OGR_Fld_Create( codec->fromUnicode( it->first ).constData(), OFTString );
-      OGR_Fld_SetWidth( field, width );
-    }
-    else if ( fields[0] == QLatin1String( "Date" ) )
-    {
-      field = OGR_Fld_Create( codec->fromUnicode( it->first ).constData(), OFTDate );
-    }
-    else if ( fields[0] == QLatin1String( "Time" ) )
-    {
-      field = OGR_Fld_Create( codec->fromUnicode( it->first ).constData(), OFTTime );
-    }
-    else if ( fields[0] == QLatin1String( "DateTime" ) )
-    {
-      field = OGR_Fld_Create( codec->fromUnicode( it->first ).constData(), OFTDateTime );
-    }
-    else
-    {
-      QgsMessageLog::logMessage( QObject::tr( "field %1 with unsupported type %2 skipped" ).arg( it->first, fields[0] ), QObject::tr( "OGR" ) );
-      continue;
-    }
-
-    if ( OGR_L_CreateField( layer, field, true ) != OGRERR_NONE )
-    {
-      QgsMessageLog::logMessage( QObject::tr( "creation of field %1 failed" ).arg( it->first ), QObject::tr( "OGR" ) );
-    }
-  }
-
-  dataSource.reset();
-
-  if ( driverName == QLatin1String( "ESRI Shapefile" ) )
-  {
-    int index = uri.indexOf( QLatin1String( ".shp" ), Qt::CaseInsensitive );
-    if ( index > 0 )
-    {
-      QString layerName = uri.left( index );
-      QFile prjFile( layerName + ".qpj" );
-      if ( prjFile.exists() )
-        prjFile.remove();
-    }
-  }
-
-  QgsDebugMsgLevel( QStringLiteral( "GDAL Version number %1" ).arg( GDAL_VERSION_NUM ), 2 );
-  if ( reference )
-  {
-    OSRRelease( reference );
-  }
-  return true;
-}
-
-QList<QgsDataItemProvider *> QgsOgrProviderMetadata::dataItemProviders() const
-{
-  QList< QgsDataItemProvider * > providers;
-  providers << new QgsOgrDataItemProvider;
-  providers << new QgsGeoPackageDataItemProvider;
-  return providers;
 }
 
 QgsCoordinateReferenceSystem QgsOgrProvider::crs() const
@@ -4673,333 +3511,6 @@ QByteArray QgsOgrProvider::quotedIdentifier( const QByteArray &field ) const
   return QgsOgrProviderUtils::quotedIdentifier( field, mGDALDriverName );
 }
 
-QString QgsOgrProviderUtils::connectionPoolId( const QString &dataSourceURI, bool shareSameDatasetAmongLayers )
-{
-  if ( shareSameDatasetAmongLayers )
-  {
-    // If the file part of the URI is really a file, then use it as the
-    // connection pool id (for example, so that all layers of a .gpkg file can
-    // use the same GDAL dataset object)
-    // Otherwise use the datasourceURI
-    // Not completely sure about this logic. But at least, for GeoPackage this
-    // works fine with multi layer datasets.
-    QString filePath = dataSourceURI.left( dataSourceURI.indexOf( QLatin1Char( '|' ) ) );
-    QFileInfo fi( filePath );
-    if ( fi.isFile() )
-      return filePath;
-  }
-  return dataSourceURI;
-}
-
-GDALDatasetH QgsOgrProviderUtils::GDALOpenWrapper( const char *pszPath, bool bUpdate, char **papszOpenOptionsIn, GDALDriverH *phDriver )
-{
-  CPLErrorReset();
-
-  char **papszOpenOptions = CSLDuplicate( papszOpenOptionsIn );
-
-  QString filePath( QString::fromUtf8( pszPath ) );
-
-  bool bIsGpkg = QFileInfo( filePath ).suffix().compare( QLatin1String( "gpkg" ), Qt::CaseInsensitive ) == 0;
-  bool bIsLocalGpkg = false;
-  if ( bIsGpkg &&
-       IsLocalFile( filePath ) &&
-       !CPLGetConfigOption( "OGR_SQLITE_JOURNAL", nullptr ) &&
-       QgsSettings().value( QStringLiteral( "qgis/walForSqlite3" ), true ).toBool() )
-  {
-    // For GeoPackage, we force opening of the file in WAL (Write Ahead Log)
-    // mode so as to avoid readers blocking writer(s), and vice-versa.
-    // https://www.sqlite.org/wal.html
-    // But only do that on a local file since WAL is advertised not to work
-    // on network shares
-    CPLSetThreadLocalConfigOption( "OGR_SQLITE_JOURNAL", "WAL" );
-    bIsLocalGpkg = true;
-  }
-  else if ( bIsGpkg )
-  {
-    // If WAL isn't set, we explicitly disable it, as it is persistent and it
-    // may have been set on a previous connection.
-    CPLSetThreadLocalConfigOption( "OGR_SQLITE_JOURNAL", "DELETE" );
-  }
-
-  bool modify_OGR_GPKG_FOREIGN_KEY_CHECK = !CPLGetConfigOption( "OGR_GPKG_FOREIGN_KEY_CHECK", nullptr );
-  if ( modify_OGR_GPKG_FOREIGN_KEY_CHECK )
-  {
-    CPLSetThreadLocalConfigOption( "OGR_GPKG_FOREIGN_KEY_CHECK", "NO" );
-  }
-
-  const int nOpenFlags = GDAL_OF_VECTOR | ( bUpdate ? GDAL_OF_UPDATE : 0 );
-  GDALDatasetH hDS = GDALOpenEx( pszPath, nOpenFlags, nullptr, papszOpenOptions, nullptr );
-  CSLDestroy( papszOpenOptions );
-
-  CPLSetThreadLocalConfigOption( "OGR_SQLITE_JOURNAL", nullptr );
-  if ( modify_OGR_GPKG_FOREIGN_KEY_CHECK )
-  {
-    CPLSetThreadLocalConfigOption( "OGR_GPKG_FOREIGN_KEY_CHECK", nullptr );
-  }
-
-  if ( !hDS )
-  {
-    if ( phDriver )
-      *phDriver = nullptr;
-    return nullptr;
-  }
-  GDALDriverH hDrv = GDALGetDatasetDriver( hDS );
-  if ( bIsLocalGpkg && strcmp( GDALGetDriverShortName( hDrv ), "GPKG" ) == 0 )
-  {
-    QMutexLocker locker( sGlobalMutex() );
-    ( *sMapCountOpenedDS() )[ filePath ]++;
-    ( *sMapDSHandleToUpdateMode() )[ hDS ] = bUpdate;
-  }
-  if ( phDriver )
-    *phDriver = hDrv;
-
-  if ( bUpdate && bIsGpkg && strcmp( GDALGetDriverShortName( hDrv ), "GPKG" ) == 0 )
-  {
-    // Explicitly enable foreign key enforcement (for new transactions)
-    GDALDatasetExecuteSQL( hDS, "PRAGMA foreign_keys = ON", nullptr, nullptr );
-
-#if GDAL_VERSION_NUM < GDAL_COMPUTE_VERSION(3, 3, 1)
-    // Fix wrong gpkg_metadata_reference_column_name_update trigger that was
-    // generated by GDAL < 2.4.0
-    OGRLayerH hSqlLyr = GDALDatasetExecuteSQL(
-                          hDS,
-                          "SELECT sql FROM sqlite_master WHERE type = 'trigger' AND "
-                          "NAME ='gpkg_metadata_reference_column_name_update' AND "
-                          "sql LIKE '%column_nameIS%'",
-                          nullptr, nullptr );
-    if ( hSqlLyr )
-    {
-      QString triggerSql;
-      gdal::ogr_feature_unique_ptr hFeat( OGR_L_GetNextFeature( hSqlLyr ) );
-      if ( hFeat )
-      {
-        triggerSql = OGR_F_GetFieldAsString( hFeat.get(), 0 );
-      }
-      GDALDatasetReleaseResultSet( hDS, hSqlLyr );
-
-      if ( !triggerSql.isEmpty() )
-      {
-        GDALDatasetExecuteSQL(
-          hDS, "DROP TRIGGER gpkg_metadata_reference_column_name_update",
-          nullptr, nullptr );
-        GDALDatasetExecuteSQL(
-          hDS,
-          triggerSql.replace( QLatin1String( "column_nameIS" ),
-                              QLatin1String( "column_name IS" ) ).toUtf8().toStdString().c_str(),
-          nullptr, nullptr );
-      }
-    }
-#endif
-  }
-
-  return hDS;
-}
-
-static bool IsLocalFile( const QString &path )
-{
-  QString dirName( QFileInfo( path ).absolutePath() );
-  // Start with the OS specific methods since the QT >= 5.4 method just
-  // return a string and not an enumerated type.
-#if defined(Q_OS_WIN)
-  if ( dirName.startsWith( "\\\\" ) || dirName.startsWith( "//" ) )
-    return false;
-  if ( dirName.length() >= 3 && dirName[1] == ':' &&
-       ( dirName[2] == '\\' || dirName[2] == '/' ) )
-  {
-    dirName.resize( 3 );
-    return GetDriveType( dirName.toLatin1().constData() ) != DRIVE_REMOTE;
-  }
-  return true;
-#elif defined(Q_OS_LINUX)
-  struct statfs sStatFS;
-  if ( statfs( dirName.toLatin1().constData(), &sStatFS ) == 0 )
-  {
-    // Codes from http://man7.org/linux/man-pages/man2/statfs.2.html
-    if ( sStatFS.f_type == 0x6969 /* NFS */ ||
-         sStatFS.f_type == 0x517b /* SMB */ ||
-         sStatFS.f_type == 0xff534d42ul /* CIFS */ ||
-         sStatFS.f_type == 0xfe534d42ul /* CIFS */ )
-    {
-      return false;
-    }
-  }
-  return true;
-#else
-  QStorageInfo info( dirName );
-  const QString fileSystem( info.fileSystemType() );
-  bool isLocal = path != QLatin1String( "nfs" ) && path != QLatin1String( "smbfs" );
-  if ( !isLocal )
-    QgsDebugMsgLevel( QStringLiteral( "Filesystem for %1 is %2" ).arg( path, fileSystem ), 2 );
-  return isLocal;
-#endif
-}
-
-void QgsOgrProviderUtils::GDALCloseWrapper( GDALDatasetH hDS )
-{
-  if ( !hDS )
-    return;
-  GDALDriverH mGDALDriver = GDALGetDatasetDriver( hDS );
-  QString mGDALDriverName = GDALGetDriverShortName( mGDALDriver );
-  QString datasetName( QString::fromUtf8( GDALGetDescription( hDS ) ) );
-  if ( mGDALDriverName == QLatin1String( "GPKG" ) &&
-       IsLocalFile( datasetName ) &&
-       !CPLGetConfigOption( "OGR_SQLITE_JOURNAL", nullptr ) )
-  {
-    bool openedAsUpdate = false;
-    bool tryReturnToWall = false;
-    {
-      QMutexLocker locker( sGlobalMutex() );
-      ( *sMapCountOpenedDS() )[ datasetName ] --;
-      if ( ( *sMapCountOpenedDS() )[ datasetName ] == 0 )
-      {
-        sMapCountOpenedDS()->remove( datasetName );
-        openedAsUpdate = ( *sMapDSHandleToUpdateMode() )[hDS];
-        tryReturnToWall = true;
-      }
-      sMapDSHandleToUpdateMode()->remove( hDS );
-    }
-    if ( tryReturnToWall )
-    {
-      bool bSuccess = false;
-      if ( openedAsUpdate )
-      {
-        // We need to reset all iterators on layers, otherwise we will not
-        // be able to change journal_mode.
-        int layerCount = GDALDatasetGetLayerCount( hDS );
-        for ( int i = 0; i < layerCount; i ++ )
-        {
-          OGR_L_ResetReading( GDALDatasetGetLayer( hDS, i ) );
-        }
-
-        CPLPushErrorHandler( CPLQuietErrorHandler );
-        QgsDebugMsgLevel( QStringLiteral( "GPKG: Trying to return to delete mode" ), 2 );
-        OGRLayerH hSqlLyr = GDALDatasetExecuteSQL( hDS,
-                            "PRAGMA journal_mode = delete",
-                            nullptr, nullptr );
-        if ( hSqlLyr )
-        {
-          gdal::ogr_feature_unique_ptr hFeat( OGR_L_GetNextFeature( hSqlLyr ) );
-          if ( hFeat )
-          {
-            const char *pszRet = OGR_F_GetFieldAsString( hFeat.get(), 0 );
-            bSuccess = EQUAL( pszRet, "delete" );
-            QgsDebugMsgLevel( QStringLiteral( "Return: %1" ).arg( pszRet ), 2 );
-          }
-        }
-        else if ( CPLGetLastErrorType() != CE_None )
-        {
-          QgsDebugMsg( QStringLiteral( "Return: %1" ).arg( CPLGetLastErrorMsg() ) );
-        }
-        GDALDatasetReleaseResultSet( hDS, hSqlLyr );
-        CPLPopErrorHandler();
-      }
-      GDALClose( hDS );
-
-      // If the file was opened in read-only mode, or if the above failed,
-      // we need to reopen it in update mode
-      if ( !bSuccess )
-      {
-        if ( openedAsUpdate )
-        {
-          QgsDebugMsgLevel( QStringLiteral( "GPKG: Trying again" ), 2 );
-        }
-        else
-        {
-          QgsDebugMsgLevel( QStringLiteral( "GPKG: Trying to return to delete mode" ), 2 );
-        }
-        CPLSetThreadLocalConfigOption( "OGR_SQLITE_JOURNAL", "DELETE" );
-        hDS = GDALOpenEx( datasetName.toUtf8().constData(), GDAL_OF_VECTOR | GDAL_OF_UPDATE, nullptr, nullptr, nullptr );
-        CPLSetThreadLocalConfigOption( "OGR_SQLITE_JOURNAL", nullptr );
-        if ( hDS )
-        {
-#ifdef QGISDEBUG
-          CPLPushErrorHandler( CPLQuietErrorHandler );
-          OGRLayerH hSqlLyr = GDALDatasetExecuteSQL( hDS,
-                              "PRAGMA journal_mode",
-                              nullptr, nullptr );
-          CPLPopErrorHandler();
-          if ( hSqlLyr != nullptr )
-          {
-            gdal::ogr_feature_unique_ptr hFeat( OGR_L_GetNextFeature( hSqlLyr ) );
-            if ( hFeat != nullptr )
-            {
-              const char *pszRet = OGR_F_GetFieldAsString( hFeat.get(), 0 );
-              QgsDebugMsgLevel( QStringLiteral( "Return: %1" ).arg( pszRet ), 2 );
-            }
-            GDALDatasetReleaseResultSet( hDS, hSqlLyr );
-          }
-#endif
-          GDALClose( hDS );
-        }
-      }
-    }
-    else
-    {
-      GDALClose( hDS );
-    }
-  }
-
-#if GDAL_VERSION_NUM >= GDAL_COMPUTE_VERSION(3,1,0) && GDAL_VERSION_NUM <= GDAL_COMPUTE_VERSION(3,1,3)
-  else if ( mGDALDriverName == QLatin1String( "XLSX" ) ||
-            mGDALDriverName == QLatin1String( "ODS" ) )
-  {
-    // Workaround bug in GDAL 3.1.0 to 3.1.3 that creates XLSX and ODS files incompatible with LibreOffice due to use of ZIP64
-    CPLSetThreadLocalConfigOption( "CPL_CREATE_ZIP64", "NO" );
-    GDALClose( hDS );
-    CPLSetThreadLocalConfigOption( "CPL_CREATE_ZIP64", nullptr );
-  }
-#endif
-
-  else
-  {
-    GDALClose( hDS );
-  }
-}
-
-QByteArray QgsOgrProviderUtils::quotedIdentifier( QByteArray field, const QString &driverName )
-{
-  if ( driverName == QLatin1String( "MySQL" ) )
-  {
-    field.replace( '\\', "\\\\" );
-    field.replace( '`', "``" );
-    return field.prepend( '`' ).append( '`' );
-  }
-  else
-  {
-    field.replace( '\\', "\\\\" );
-    field.replace( '"', "\\\"" );
-    field.replace( '\'', "\\'" );
-    return field.prepend( '\"' ).append( '\"' );
-  }
-}
-
-QString QgsOgrProviderUtils::quotedValue( const QVariant &value )
-{
-  if ( value.isNull() )
-    return QStringLiteral( "NULL" );
-
-  switch ( value.type() )
-  {
-    case QVariant::Int:
-    case QVariant::LongLong:
-    case QVariant::Double:
-      return value.toString();
-
-    case QVariant::Bool:
-      //OGR does not support boolean literals
-      return value.toBool() ? "1" : "0";
-
-    default:
-    case QVariant::String:
-      QString v = value.toString();
-      v.replace( '\'', QLatin1String( "''" ) );
-      if ( v.contains( '\\' ) )
-        return v.replace( '\\', QLatin1String( "\\\\" ) ).prepend( "E'" ).append( '\'' );
-      else
-        return v.prepend( '\'' ).append( '\'' );
-  }
-}
-
 bool QgsOgrProvider::syncToDisc()
 {
   QgsOgrConnPool::instance()->invalidateConnections( QgsOgrProviderUtils::connectionPoolId( dataSourceUri( true ), mShareSameDatasetAmongLayers ) );
@@ -5132,42 +3643,6 @@ OGRwkbGeometryType QgsOgrProvider::ogrWkbSingleFlatten( OGRwkbGeometryType type 
   }
 }
 
-OGRLayerH QgsOgrProviderUtils::setSubsetString( OGRLayerH layer, GDALDatasetH ds, QTextCodec *encoding, const QString &subsetString )
-{
-  QByteArray layerName = OGR_FD_GetName( OGR_L_GetLayerDefn( layer ) );
-  GDALDriverH driver = GDALGetDatasetDriver( ds );
-  QString driverName = GDALGetDriverShortName( driver );
-
-  if ( driverName == QLatin1String( "ODBC" ) ) //the odbc driver does not like schema names for subset
-  {
-    QString layerNameString = encoding->toUnicode( layerName );
-    int dotIndex = layerNameString.indexOf( '.' );
-    if ( dotIndex > 1 )
-    {
-      QString modifiedLayerName = layerNameString.right( layerNameString.size() - dotIndex - 1 );
-      layerName = encoding->fromUnicode( modifiedLayerName );
-    }
-  }
-  OGRLayerH subsetLayer = nullptr;
-  if ( subsetString.startsWith( QLatin1String( "SELECT " ), Qt::CaseInsensitive ) )
-  {
-    QByteArray sql = encoding->fromUnicode( subsetString );
-
-    QgsDebugMsgLevel( QStringLiteral( "SQL: %1" ).arg( encoding->toUnicode( sql ) ), 2 );
-    subsetLayer = GDALDatasetExecuteSQL( ds, sql.constData(), nullptr, nullptr );
-  }
-  else
-  {
-    if ( OGR_L_SetAttributeFilter( layer, encoding->fromUnicode( subsetString ).constData() ) != OGRERR_NONE )
-    {
-      return nullptr;
-    }
-    subsetLayer = layer;
-  }
-
-  return subsetLayer;
-}
-
 void QgsOgrProvider::open( OpenMode mode )
 {
   bool openReadOnly = false;
@@ -5254,9 +3729,9 @@ void QgsOgrProvider::open( OpenMode mode )
 #if GDAL_VERSION_NUM >= GDAL_COMPUTE_VERSION(3,1,0)
     // assume trusted data to get more speed
     if ( mGDALDriverName == QLatin1String( "FlatGeobuf" ) &&
-         !options.contains( "VERIFY_BUFFERS=YES" ) )
+         !options.contains( QStringLiteral( "VERIFY_BUFFERS=YES" ) ) )
     {
-      options << "VERIFY_BUFFERS=NO";
+      options << QStringLiteral( "VERIFY_BUFFERS=NO" );
     }
 #endif
 
@@ -5532,2218 +4007,4 @@ QString QgsOgrProvider::fileVectorFilters() const
   return QgsOgrProviderUtils::fileVectorFilters();
 }
 
-QString QgsOgrProviderUtils::DatasetIdentification::toString() const
-{
-  return dsName +
-         ( updateMode ?
-           QStringLiteral( "update" ) : QStringLiteral( "read-only" ) ) +
-         options.join( ',' );
-}
-
-bool QgsOgrProviderUtils::DatasetIdentification::operator<
-( const QgsOgrProviderUtils::DatasetIdentification &other ) const
-{
-  return toString() < other.toString();
-}
-
-static GDALDatasetH OpenHelper( const QString &dsName,
-                                bool updateMode,
-                                const QStringList &options )
-{
-  char **papszOpenOptions = nullptr;
-  const auto constOptions = options;
-  for ( QString option : constOptions )
-  {
-    papszOpenOptions = CSLAddString( papszOpenOptions,
-                                     option.toUtf8().constData() );
-  }
-  GDALDatasetH hDS = QgsOgrProviderUtils::GDALOpenWrapper(
-                       QgsOgrProviderUtils::expandAuthConfig( dsName ).toUtf8().constData(), updateMode, papszOpenOptions, nullptr );
-  CSLDestroy( papszOpenOptions );
-  return hDS;
-}
-
-void QgsOgrProviderUtils::invalidateCachedDatasets( const QString &dsName )
-{
-  QMutexLocker locker( sGlobalMutex() );
-  while ( true )
-  {
-    bool erased = false;
-    for ( auto iter = sMapSharedDS.begin(); iter != sMapSharedDS.end(); ++iter )
-    {
-      if ( iter.key().dsName == dsName )
-      {
-        sMapSharedDS.erase( iter );
-        erased = true;
-        break;
-      }
-    }
-    if ( !erased )
-      break;
-  }
-}
-
-
-QgsOgrDatasetSharedPtr QgsOgrProviderUtils::getAlreadyOpenedDataset( const QString &dsName )
-{
-  QMutexLocker locker( sGlobalMutex() );
-  for ( auto iter = sMapSharedDS.begin(); iter != sMapSharedDS.end(); ++iter )
-  {
-    auto ident = iter.key();
-    if ( ident.dsName == dsName && ident.updateMode )
-    {
-      // Browse through this list, to look for the first available DatasetWithLayers*
-      // instance that is in update mode (hoping there's only one...)
-      auto &datasetList = iter.value();
-      for ( const auto &ds : datasetList )
-      {
-        Q_ASSERT( ds->refCount > 0 );
-        return QgsOgrDataset::create( ident, ds );
-      }
-    }
-  }
-  return nullptr;
-}
-
-
-QgsOgrLayerUniquePtr QgsOgrProviderUtils::getLayer( const QString &dsName,
-    int layerIndex,
-    QString &errCause )
-{
-  QMutexLocker locker( sGlobalMutex() );
-  for ( auto iter = sMapSharedDS.begin(); iter != sMapSharedDS.end(); ++iter )
-  {
-    if ( iter.key().dsName == dsName )
-    {
-      // Browse through this list, to look for a DatasetWithLayers*
-      // instance that don't use yet our layer of interest
-      auto &datasetList = iter.value();
-      const auto constDatasetList = datasetList;
-      for ( QgsOgrProviderUtils::DatasetWithLayers *ds : constDatasetList )
-      {
-        if ( !ds->canBeShared )
-          continue;
-        Q_ASSERT( ds->refCount > 0 );
-
-        QString layerName;
-        OGRLayerH hLayer;
-        {
-          QMutexLocker lockerDS( &ds->mutex );
-          hLayer = GDALDatasetGetLayer(
-                     ds->hDS, layerIndex );
-          if ( hLayer )
-          {
-            layerName = QString::fromUtf8( OGR_L_GetName( hLayer ) );
-          }
-        }
-        if ( !hLayer )
-        {
-          errCause = QObject::tr( "Cannot find layer %1." ).arg( layerIndex );
-          return nullptr;
-        }
-        return getLayer( dsName, iter.key().updateMode, iter.key().options, layerName, errCause, true );
-      }
-    }
-  }
-  return getLayer( dsName, false, QStringList(), layerIndex, errCause, true );
-}
-
-QgsOgrLayerUniquePtr QgsOgrProviderUtils::getLayer( const QString &dsName,
-    bool updateMode,
-    const QStringList &options,
-    int layerIndex,
-    QString &errCause,
-    bool checkModificationDateAgainstCache )
-{
-  QMutexLocker locker( sGlobalMutex() );
-
-  // The idea is that we want to minimize the number of GDALDatasetH
-  // handles openeded. But we have constraints. We do not want that 2
-  // callers of getLayer() with the same input parameters get the same
-  // GDALDatasetH since iterators over features of that layer would conflict
-
-  QgsOgrProviderUtils::DatasetIdentification ident;
-  ident.dsName = dsName;
-  ident.updateMode = updateMode;
-  ident.options = options;
-  // Find if there's a list of DatasetWithLayers* that match our
-  // (dsName, updateMode, options) criteria
-  auto iter = sMapSharedDS.find( ident );
-  if ( iter != sMapSharedDS.end() )
-  {
-    // Browse through this list, to look for a DatasetWithLayers*
-    // instance that don't use yet our layer of interest
-    auto datasetList = iter.value();
-    const auto constDatasetList = datasetList;
-    for ( QgsOgrProviderUtils::DatasetWithLayers *ds : constDatasetList )
-    {
-      if ( !ds->canBeShared )
-        continue;
-      Q_ASSERT( ds->refCount > 0 );
-
-      QString layerName;
-      OGRLayerH hLayer;
-      {
-        QMutexLocker lockerDS( &ds->mutex );
-        hLayer = GDALDatasetGetLayer(
-                   ds->hDS, layerIndex );
-        if ( hLayer )
-        {
-          layerName = QString::fromUtf8( OGR_L_GetName( hLayer ) );
-        }
-      }
-      if ( !hLayer )
-      {
-        errCause = QObject::tr( "Cannot find layer %1." ).arg( layerIndex );
-        return nullptr;
-      }
-      return getLayer( dsName, updateMode, options, layerName, errCause,
-                       checkModificationDateAgainstCache );
-    }
-  }
-
-  GDALDatasetH hDS = OpenHelper( dsName, updateMode, options );
-  if ( !hDS )
-  {
-    errCause = QObject::tr( "Cannot open %1." ).arg( dsName );
-    return nullptr;
-  }
-
-  OGRLayerH hLayer = GDALDatasetGetLayer(
-                       hDS, layerIndex );
-  if ( !hLayer )
-  {
-    QgsOgrProviderUtils::GDALCloseWrapper( hDS );
-    errCause = QObject::tr( "Cannot find layer %1." ).arg( layerIndex );
-    return nullptr;
-  }
-  QString layerName = QString::fromUtf8( OGR_L_GetName( hLayer ) );
-
-  QgsOgrProviderUtils::DatasetWithLayers *ds =
-    new QgsOgrProviderUtils::DatasetWithLayers;
-  ds->hDS = hDS;
-
-  GDALDriverH driver = GDALGetDatasetDriver( hDS );
-  QString driverName = GDALGetDriverShortName( driver );
-  ds->canBeShared = canDriverShareSameDatasetAmongLayers( driverName, updateMode, dsName );
-
-  QgsOgrLayerUniquePtr layer = QgsOgrLayer::CreateForLayer(
-                                 ident, layerName, ds, hLayer );
-  ds->setLayers[layerName] = layer.get();
-
-  QList<DatasetWithLayers *> datasetList;
-  datasetList.push_back( ds );
-  sMapSharedDS[ident] = datasetList;
-
-  return layer;
-}
-
-QgsOgrLayerUniquePtr QgsOgrProviderUtils::getLayer( const QString &dsName,
-    const QString &layerName,
-    QString &errCause )
-{
-  QMutexLocker locker( sGlobalMutex() );
-
-  for ( auto iter = sMapSharedDS.begin(); iter != sMapSharedDS.end(); ++iter )
-  {
-    if ( iter.key().dsName == dsName )
-    {
-      // Browse through this list, to look for a DatasetWithLayers*
-      // instance that don't use yet our layer of interest
-      auto &datasetList = iter.value();
-      const auto constDatasetList = datasetList;
-      for ( QgsOgrProviderUtils::DatasetWithLayers *ds : constDatasetList )
-      {
-        if ( !ds->canBeShared )
-          continue;
-        Q_ASSERT( ds->refCount > 0 );
-
-        auto iter2 = ds->setLayers.find( layerName );
-        if ( iter2 == ds->setLayers.end() )
-        {
-          OGRLayerH hLayer;
-          {
-            QMutexLocker lockerDS( &ds->mutex );
-            hLayer = GDALDatasetGetLayerByName(
-                       ds->hDS, layerName.toUtf8().constData() );
-          }
-          if ( !hLayer )
-          {
-            // Shouldn't happen really !
-            errCause = QObject::tr( "Cannot find layer %1." ).arg( layerName );
-            return nullptr;
-          }
-          OGR_L_SetAttributeFilter( hLayer, nullptr );
-
-          QgsOgrLayerUniquePtr layer = QgsOgrLayer::CreateForLayer(
-                                         iter.key(), layerName, ds, hLayer );
-          ds->setLayers[layerName] = layer.get();
-          return layer;
-        }
-      }
-    }
-  }
-  return getLayer( dsName, false, QStringList(), layerName, errCause, true );
-}
-
-static QDateTime getLastModified( const QString &dsName )
-{
-  if ( dsName.endsWith( ".gpkg", Qt::CaseInsensitive ) )
-  {
-    QFileInfo info( dsName + "-wal" );
-    if ( info.exists() )
-      return info.lastModified();
-  }
-  return QFileInfo( dsName ).lastModified();
-}
-
-// In case we do very fast structural changes within the same second,
-// the last modified date might not change enough, so artificially
-// decrement the cache modified date, so that the file appears newer to it
-void QgsOgrProviderUtils::invalidateCachedLastModifiedDate( const QString &dsName )
-{
-  QMutexLocker locker( sGlobalMutex() );
-
-  auto iter = sMapDSNameToLastModifiedDate()->find( dsName );
-  if ( iter != sMapDSNameToLastModifiedDate()->end() )
-  {
-    QgsDebugMsgLevel( QStringLiteral( "invalidating last modified date for %1" ).arg( dsName ), 2 );
-    iter.value() = iter.value().addSecs( -10 );
-  }
-}
-
-OGRwkbGeometryType QgsOgrProviderUtils::ogrTypeFromQgisType( QgsWkbTypes::Type type )
-{
-  switch ( type )
-  {
-    case QgsWkbTypes::Point:
-      return wkbPoint;
-    case QgsWkbTypes::Point25D:
-    case QgsWkbTypes::PointZ:
-      return wkbPoint25D;
-    case QgsWkbTypes::PointM:
-      return wkbPointM;
-    case QgsWkbTypes::PointZM:
-      return wkbPointZM;
-
-    case QgsWkbTypes::LineString:
-      return wkbLineString;
-    case QgsWkbTypes::LineString25D:
-    case QgsWkbTypes::LineStringZ:
-      return wkbLineString25D;
-    case QgsWkbTypes::LineStringM:
-      return wkbLineStringM;
-    case QgsWkbTypes::LineStringZM:
-      return wkbLineStringZM;
-
-    case QgsWkbTypes::Polygon:
-      return wkbPolygon;
-    case QgsWkbTypes::Polygon25D:
-    case QgsWkbTypes::PolygonZ:
-      return wkbPolygon25D;
-    case QgsWkbTypes::PolygonM:
-      return wkbPolygonM;
-    case QgsWkbTypes::PolygonZM:
-      return wkbPolygonZM;
-
-    case QgsWkbTypes::MultiPoint:
-      return wkbMultiPoint;
-    case QgsWkbTypes::MultiPoint25D:
-    case QgsWkbTypes::MultiPointZ:
-      return wkbMultiPoint25D;
-    case QgsWkbTypes::MultiPointM:
-      return wkbMultiPointM;
-    case QgsWkbTypes::MultiPointZM:
-      return wkbMultiPointZM;
-
-    case QgsWkbTypes::MultiLineString:
-      return wkbMultiLineString;
-    case QgsWkbTypes::MultiLineString25D:
-    case QgsWkbTypes::MultiLineStringZ:
-      return wkbMultiLineString25D;
-    case QgsWkbTypes::MultiLineStringM:
-      return wkbMultiLineStringM;
-    case QgsWkbTypes::MultiLineStringZM:
-      return wkbMultiLineStringZM;
-
-    case QgsWkbTypes::MultiPolygon:
-      return wkbMultiPolygon;
-    case QgsWkbTypes::MultiPolygon25D:
-    case QgsWkbTypes::MultiPolygonZ:
-      return wkbMultiPolygon25D;
-    case QgsWkbTypes::MultiPolygonM:
-      return wkbMultiPolygonM;
-    case QgsWkbTypes::MultiPolygonZM:
-      return wkbMultiPolygonZM;
-
-    case QgsWkbTypes::CircularString:
-      return wkbCircularString;
-    case QgsWkbTypes::CircularStringZ:
-      return wkbCircularStringZ;
-    case QgsWkbTypes::CircularStringM:
-      return wkbCircularStringM;
-    case QgsWkbTypes::CircularStringZM:
-      return wkbCircularStringZM;
-
-    case QgsWkbTypes::CompoundCurve:
-      return wkbCompoundCurve;
-    case QgsWkbTypes::CompoundCurveZ:
-      return wkbCompoundCurveZ;
-    case QgsWkbTypes::CompoundCurveM:
-      return wkbCompoundCurveM;
-    case QgsWkbTypes::CompoundCurveZM:
-      return wkbCompoundCurveZM;
-
-    case QgsWkbTypes::CurvePolygon:
-      return wkbCurvePolygon;
-    case QgsWkbTypes::CurvePolygonZ:
-      return wkbCurvePolygonZ;
-    case QgsWkbTypes::CurvePolygonM:
-      return wkbCurvePolygonM;
-    case QgsWkbTypes::CurvePolygonZM:
-      return wkbCurvePolygonZM;
-
-    case QgsWkbTypes::MultiCurve:
-      return wkbMultiCurve;
-    case QgsWkbTypes::MultiCurveZ:
-      return wkbMultiCurveZ;
-    case QgsWkbTypes::MultiCurveM:
-      return wkbMultiCurveM;
-    case QgsWkbTypes::MultiCurveZM:
-      return wkbMultiCurveZM;
-
-    case QgsWkbTypes::MultiSurface:
-      return wkbMultiSurface;
-    case QgsWkbTypes::MultiSurfaceZ:
-      return wkbMultiSurfaceZ;
-    case QgsWkbTypes::MultiSurfaceM:
-      return wkbMultiSurfaceM;
-    case QgsWkbTypes::MultiSurfaceZM:
-      return wkbMultiSurfaceZM;
-
-    case QgsWkbTypes::Triangle:
-      return wkbTriangle;
-    case QgsWkbTypes::TriangleZ:
-      return wkbTriangleZ;
-    case QgsWkbTypes::TriangleM:
-      return wkbTriangleM;
-    case QgsWkbTypes::TriangleZM:
-      return wkbTriangleZM;
-
-    case QgsWkbTypes::NoGeometry:
-      return wkbNone;
-
-    case QgsWkbTypes::GeometryCollection:
-      return wkbGeometryCollection;
-    case QgsWkbTypes::GeometryCollectionZ:
-      return wkbGeometryCollection25D;
-    case QgsWkbTypes::GeometryCollectionM:
-      return wkbGeometryCollectionM;
-    case QgsWkbTypes::GeometryCollectionZM:
-      return wkbGeometryCollectionZM;
-
-    case QgsWkbTypes::Unknown:
-      return wkbUnknown;
-  }
-  // no warnings!
-  return wkbUnknown;
-}
-
-QgsWkbTypes::Type QgsOgrProviderUtils::qgisTypeFromOgrType( OGRwkbGeometryType type )
-{
-  switch ( type )
-  {
-    case wkbUnknown:
-      return QgsWkbTypes::Unknown;
-    case wkbPoint:
-      return QgsWkbTypes::Point;
-    case wkbLineString:
-      return QgsWkbTypes::LineString;
-    case wkbPolygon:
-      return QgsWkbTypes::Polygon;
-    case wkbMultiPoint:
-      return QgsWkbTypes::MultiPoint;
-    case wkbMultiLineString:
-      return QgsWkbTypes::MultiLineString;
-    case wkbMultiPolygon:
-      return QgsWkbTypes::MultiPolygon;
-    case wkbGeometryCollection:
-      return QgsWkbTypes::GeometryCollection;
-    case wkbCircularString:
-      return QgsWkbTypes::CircularString;
-    case wkbCompoundCurve:
-      return QgsWkbTypes::CompoundCurve;
-    case wkbCurvePolygon:
-      return QgsWkbTypes::CurvePolygon;
-    case wkbMultiCurve:
-      return QgsWkbTypes::MultiCurve;
-    case wkbMultiSurface:
-      return QgsWkbTypes::MultiSurface;
-    case wkbTriangle:
-      return QgsWkbTypes::Triangle;
-    case wkbNone:
-      return QgsWkbTypes::NoGeometry;
-
-    case wkbCircularStringZ:
-      return QgsWkbTypes::CircularStringZ;
-    case wkbCompoundCurveZ:
-      return QgsWkbTypes::CompoundCurveZ;
-    case wkbCurvePolygonZ:
-      return QgsWkbTypes::PolygonZ;
-    case wkbMultiCurveZ:
-      return QgsWkbTypes::MultiCurveZ;
-    case wkbMultiSurfaceZ:
-      return QgsWkbTypes::MultiSurfaceZ;
-    case wkbTriangleZ:
-      return QgsWkbTypes::TriangleZ;
-
-    case wkbPointM:
-      return QgsWkbTypes::PointM;
-    case wkbLineStringM:
-      return QgsWkbTypes::LineStringM;
-    case wkbPolygonM:
-      return QgsWkbTypes::PolygonM;
-    case wkbMultiPointM:
-      return QgsWkbTypes::PointM;
-    case wkbMultiLineStringM:
-      return QgsWkbTypes::LineStringM;
-    case wkbMultiPolygonM:
-      return QgsWkbTypes::PolygonM;
-    case wkbGeometryCollectionM:
-      return QgsWkbTypes::GeometryCollectionM;
-    case wkbCircularStringM:
-      return QgsWkbTypes::CircularStringM;
-    case wkbCompoundCurveM:
-      return QgsWkbTypes::CompoundCurveM;
-    case wkbCurvePolygonM:
-      return QgsWkbTypes::PolygonM;
-    case wkbMultiCurveM:
-      return QgsWkbTypes::MultiCurveM;
-    case wkbMultiSurfaceM:
-      return QgsWkbTypes::MultiSurfaceM;
-    case wkbTriangleM:
-      return QgsWkbTypes::TriangleM;
-
-    case wkbPointZM:
-      return QgsWkbTypes::PointZM;
-    case wkbLineStringZM:
-      return QgsWkbTypes::LineStringZM;
-    case wkbPolygonZM:
-      return QgsWkbTypes::PolygonZM;
-    case wkbMultiPointZM:
-      return QgsWkbTypes::MultiPointZM;
-    case wkbMultiLineStringZM:
-      return QgsWkbTypes::MultiLineStringZM;
-    case wkbMultiPolygonZM:
-      return QgsWkbTypes::MultiPolygonZM;
-    case wkbGeometryCollectionZM:
-      return QgsWkbTypes::GeometryCollectionZM;
-    case wkbCircularStringZM:
-      return QgsWkbTypes::CircularStringZM;
-    case wkbCompoundCurveZM:
-      return QgsWkbTypes::CompoundCurveZM;
-    case wkbCurvePolygonZM:
-      return QgsWkbTypes::CurvePolygonZM;
-    case wkbMultiCurveZM:
-      return QgsWkbTypes::MultiCurveZM;
-    case wkbMultiSurfaceZM:
-      return QgsWkbTypes::MultiSurfaceZM;
-    case wkbTriangleZM:
-      return QgsWkbTypes::TriangleZM;
-
-    case wkbPoint25D:
-      return QgsWkbTypes::Point25D;
-    case wkbLineString25D:
-      return QgsWkbTypes::LineString25D;
-    case wkbPolygon25D:
-      return QgsWkbTypes::Polygon25D;
-    case wkbMultiPoint25D:
-      return QgsWkbTypes::MultiPoint25D;
-    case wkbMultiLineString25D:
-      return QgsWkbTypes::MultiLineString25D;
-    case wkbMultiPolygon25D:
-      return QgsWkbTypes::MultiPolygon25D;
-    case wkbGeometryCollection25D:
-      return QgsWkbTypes::GeometryCollectionZ;
-
-    case wkbCurve:
-    case wkbSurface:
-    case wkbCurveZ:
-    case wkbSurfaceZ:
-    case wkbCurveM:
-    case wkbSurfaceM:
-    case wkbCurveZM:
-    case wkbSurfaceZM:
-      return QgsWkbTypes::Unknown; // abstract types - no direct mapping to QGIS types
-
-    case wkbLinearRing:
-    case wkbTIN:
-    case wkbTINZ:
-    case wkbTINM:
-    case wkbTINZM:
-    case wkbPolyhedralSurface:
-    case wkbPolyhedralSurfaceZ:
-    case wkbPolyhedralSurfaceM:
-    case wkbPolyhedralSurfaceZM:
-      return QgsWkbTypes::Unknown; // unsupported types
-  }
-  return QgsWkbTypes::Unknown;
-}
-
-
-QString QgsOgrProviderUtils::expandAuthConfig( const QString &dsName )
-{
-  QString uri( dsName );
-  // Check for authcfg
-  QRegularExpression authcfgRe( " authcfg='([^']+)'" );
-  QRegularExpressionMatch match;
-  if ( uri.contains( authcfgRe, &match ) )
-  {
-    uri = uri.replace( match.captured( 0 ), QString() );
-    QString configId( match.captured( 1 ) );
-    QStringList connectionItems;
-    connectionItems << uri;
-    if ( QgsApplication::authManager()->updateDataSourceUriItems( connectionItems, configId, QStringLiteral( "ogr" ) ) )
-    {
-      uri = connectionItems.first( );
-    }
-  }
-  return uri;
-}
-
-// Must be called under the globalMutex
-bool QgsOgrProviderUtils::canUseOpenedDatasets( const QString &dsName )
-{
-  auto iter = sMapDSNameToLastModifiedDate()->find( dsName );
-  if ( iter == sMapDSNameToLastModifiedDate()->end() )
-    return true;
-  return getLastModified( dsName ) <= iter.value();
-}
-
-QgsOgrProviderUtils::DatasetWithLayers *QgsOgrProviderUtils::createDatasetWithLayers(
-  const QString &dsName,
-  bool updateMode,
-  const QStringList &options,
-  const QString &layerName,
-  const DatasetIdentification &ident,
-  QgsOgrLayerUniquePtr &layer,
-  QString &errCause )
-{
-  GDALDatasetH hDS = OpenHelper( dsName, updateMode, options );
-  if ( !hDS )
-  {
-    errCause = QObject::tr( "Cannot open %1." ).arg( dsName );
-    return nullptr;
-  }
-  ( *sMapDSNameToLastModifiedDate() )[dsName] = getLastModified( dsName );
-
-  OGRLayerH hLayer = GDALDatasetGetLayerByName(
-                       hDS, layerName.toUtf8().constData() );
-  if ( !hLayer )
-  {
-    errCause = QObject::tr( "Cannot find layer %1." ).arg( layerName );
-    QgsOgrProviderUtils::GDALCloseWrapper( hDS );
-    return nullptr;
-  }
-
-  QgsOgrProviderUtils::DatasetWithLayers *ds =
-    new QgsOgrProviderUtils::DatasetWithLayers;
-  ds->hDS = hDS;
-
-  GDALDriverH driver = GDALGetDatasetDriver( hDS );
-  QString driverName = GDALGetDriverShortName( driver );
-  ds->canBeShared = canDriverShareSameDatasetAmongLayers( driverName, updateMode, dsName );
-
-  layer = QgsOgrLayer::CreateForLayer(
-            ident, layerName, ds, hLayer );
-  ds->setLayers[layerName] = layer.get();
-  return ds;
-}
-
-QgsOgrLayerUniquePtr QgsOgrProviderUtils::getLayer( const QString &dsName,
-    bool updateMode,
-    const QStringList &options,
-    const QString &layerName,
-    QString &errCause,
-    bool checkModificationDateAgainstCache )
-{
-  QMutexLocker locker( sGlobalMutex() );
-
-  // The idea is that we want to minimize the number of GDALDatasetH
-  // handles openeded. But we have constraints. We do not want that 2
-  // callers of getLayer() with the same input parameters get the same
-  // GDALDatasetH since iterators over features of that layer would conflict
-
-  QgsOgrProviderUtils::DatasetIdentification ident;
-  ident.dsName = dsName;
-  ident.updateMode = updateMode;
-  ident.options = options;
-  // Find if there's a list of DatasetWithLayers* that match our
-  // (dsName, updateMode, options) criteria
-  auto iter = sMapSharedDS.find( ident );
-  if ( iter != sMapSharedDS.end() )
-  {
-    if ( checkModificationDateAgainstCache && !canUseOpenedDatasets( dsName ) )
-    {
-      QgsDebugMsg( QStringLiteral( "Cannot reuse existing opened dataset(s) on %1 since it has been modified" ).arg( dsName ) );
-      invalidateCachedDatasets( dsName );
-      iter = sMapSharedDS.find( ident );
-      Q_ASSERT( iter == sMapSharedDS.end() );
-    }
-  }
-  if ( iter != sMapSharedDS.end() )
-  {
-    // Browse through this list, to look for a DatasetWithLayers*
-    // instance that don't use yet our layer of interest
-    auto &datasetList = iter.value();
-    const auto constDatasetList = datasetList;
-    for ( QgsOgrProviderUtils::DatasetWithLayers *ds : constDatasetList )
-    {
-      if ( !ds->canBeShared )
-        continue;
-      Q_ASSERT( ds->refCount > 0 );
-
-      auto iter2 = ds->setLayers.find( layerName );
-      if ( iter2 == ds->setLayers.end() )
-      {
-        OGRLayerH hLayer;
-        {
-          QMutexLocker lockerDS( &ds->mutex );
-          hLayer = GDALDatasetGetLayerByName(
-                     ds->hDS, layerName.toUtf8().constData() );
-        }
-        if ( !hLayer )
-        {
-          // Shouldn't happen really !
-          errCause = QObject::tr( "Cannot find layer %1." ).arg( layerName );
-          return nullptr;
-        }
-        OGR_L_SetAttributeFilter( hLayer, nullptr );
-
-        QgsOgrLayerUniquePtr layer = QgsOgrLayer::CreateForLayer(
-                                       ident, layerName, ds, hLayer );
-        ds->setLayers[layerName] = layer.get();
-        return layer;
-      }
-    }
-
-    // All existing DatasetWithLayers* already reference our layer of
-    // interest, so instantiate a new DatasetWithLayers*
-    QgsOgrLayerUniquePtr layer;
-    QgsOgrProviderUtils::DatasetWithLayers *ds =
-      createDatasetWithLayers( dsName, updateMode, options, layerName, ident, layer, errCause );
-    if ( !ds )
-      return nullptr;
-
-    datasetList.push_back( ds );
-
-    return layer;
-  }
-
-  QgsOgrLayerUniquePtr layer;
-  QgsOgrProviderUtils::DatasetWithLayers *ds =
-    createDatasetWithLayers( dsName, updateMode, options, layerName, ident, layer, errCause );
-  if ( !ds )
-    return nullptr;
-
-  QList<DatasetWithLayers *> datasetList;
-  datasetList.push_back( ds );
-  sMapSharedDS[ident] = datasetList;
-
-  return layer;
-}
-
-QgsOgrLayerUniquePtr QgsOgrProviderUtils::getSqlLayer( QgsOgrLayer *baseLayer,
-    OGRLayerH hSqlLayer,
-    const QString &sql )
-{
-  QgsOgrProviderUtils::DatasetIdentification ident;
-  ident.dsName = baseLayer->datasetName();
-  ident.updateMode = baseLayer->updateMode();
-  ident.options = baseLayer->options();
-  return QgsOgrLayer::CreateForSql( ident, sql, baseLayer->ds, hSqlLayer );
-}
-
-void QgsOgrProviderUtils::releaseInternal( const DatasetIdentification &ident,
-    DatasetWithLayers *ds,
-    bool removeFromDatasetList )
-{
-
-  ds->refCount --;
-  if ( ds->refCount == 0 )
-  {
-    Q_ASSERT( ds->setLayers.isEmpty() );
-
-    if ( removeFromDatasetList )
-    {
-      auto iter = sMapSharedDS.find( ident );
-      if ( iter != sMapSharedDS.end() )
-      {
-        auto &datasetList = iter.value();
-        int i = 0;
-
-        // Normally there should be a match, except for datasets that
-        // have been invalidated
-        const auto constDatasetList = datasetList;
-        for ( QgsOgrProviderUtils::DatasetWithLayers *dsIter : constDatasetList )
-        {
-          if ( dsIter == ds )
-          {
-            datasetList.removeAt( i );
-            break;
-          }
-          i ++;
-        }
-
-        if ( datasetList.isEmpty() )
-          sMapSharedDS.erase( iter );
-      }
-    }
-    QgsOgrProviderUtils::GDALCloseWrapper( ds->hDS );
-    delete ds;
-  }
-}
-
-void QgsOgrProviderUtils::release( QgsOgrLayer *&layer )
-{
-  if ( !layer )
-    return;
-
-  QMutexLocker locker( sGlobalMutex() );
-
-  if ( !layer->isSqlLayer )
-  {
-    layer->ds->setLayers.remove( layer->layerName );
-  }
-  else
-  {
-    QMutexLocker lockerDS( &layer->ds->mutex );
-    GDALDatasetReleaseResultSet( layer->ds->hDS, layer->hLayer );
-  }
-
-  releaseInternal( layer->ident, layer->ds,  !layer->isSqlLayer );
-
-  delete layer;
-  layer = nullptr;
-}
-
-
-void QgsOgrProviderUtils::releaseDataset( QgsOgrDataset *&ds )
-{
-  if ( !ds )
-    return;
-
-  QMutexLocker locker( sGlobalMutex() );
-  releaseInternal( ds->mIdent, ds->mDs, true );
-  delete ds;
-  ds = nullptr;
-}
-
-bool QgsOgrProviderUtils::canDriverShareSameDatasetAmongLayers( const QString &driverName )
-{
-  return driverName != QLatin1String( "OSM" );
-}
-
-bool QgsOgrProviderUtils::canDriverShareSameDatasetAmongLayers( const QString &driverName,
-    bool updateMode,
-    const QString &dsName )
-{
-  // For .shp.zip with multiple layers, in update mode, we want that each
-  // layer has its own dataset, so that when its gets closed and reopened,
-  // the .shp.zip is updated. Otherwise if we share the same dataset, the .shp.zip
-  // would only be updated when all layers are unloaded, and thus readers will see
-  // an outdated version of the .shp.zip. This works only if editing operations are
-  // done separately on layers (which is how it works from the GUI)
-  return canDriverShareSameDatasetAmongLayers( driverName ) &&
-         !( updateMode && dsName.endsWith( QLatin1String( ".shp.zip" ), Qt::CaseInsensitive ) );
-}
-
-
-QgsOgrDatasetSharedPtr QgsOgrDataset::create( const QgsOgrProviderUtils::DatasetIdentification &ident,
-    QgsOgrProviderUtils::DatasetWithLayers *ds )
-{
-  QgsOgrDatasetSharedPtr dsRet = QgsOgrDatasetSharedPtr( new QgsOgrDataset(), QgsOgrProviderUtils::releaseDataset );
-  dsRet->mIdent = ident;
-  dsRet->mDs = ds;
-  dsRet->mDs->refCount ++;
-  return dsRet;
-}
-
-bool QgsOgrDataset::executeSQLNoReturn( const QString &sql )
-{
-  QMutexLocker locker( &mutex() );
-  CPLErrorReset();
-  OGRLayerH hSqlLayer = GDALDatasetExecuteSQL(
-                          mDs->hDS, sql.toUtf8().constData(), nullptr, nullptr );
-  bool ret = CPLGetLastErrorType() == CE_None;
-  GDALDatasetReleaseResultSet( mDs->hDS, hSqlLayer );
-  return ret;
-}
-
-
-OGRLayerH QgsOgrDataset::getLayerFromNameOrIndex( const QString &layerName, int layerIndex )
-{
-  QMutexLocker locker( &mutex() );
-
-  OGRLayerH layer;
-  if ( !layerName.isEmpty() )
-  {
-    layer = GDALDatasetGetLayerByName( mDs->hDS, layerName.toUtf8().constData() );
-  }
-  else
-  {
-    layer = GDALDatasetGetLayer( mDs->hDS, layerIndex );
-  }
-  return layer;
-}
-
-void QgsOgrDataset::releaseResultSet( OGRLayerH hSqlLayer )
-{
-  QMutexLocker locker( &mutex() );
-  GDALDatasetReleaseResultSet( mDs->hDS, hSqlLayer );
-}
-
-QgsOgrLayer::QgsOgrLayer()
-{
-  oFDefn.layer = this;
-}
-
-QgsOgrLayerUniquePtr QgsOgrLayer::CreateForLayer(
-  const QgsOgrProviderUtils::DatasetIdentification &ident,
-  const QString &layerName,
-  QgsOgrProviderUtils::DatasetWithLayers *ds,
-  OGRLayerH hLayer )
-{
-  QgsOgrLayerUniquePtr layer( new QgsOgrLayer() );
-  layer->ident = ident;
-  layer->isSqlLayer = false;
-  layer->layerName = layerName;
-  layer->ds = ds;
-  layer->hLayer = hLayer;
-  {
-    QMutexLocker locker( &ds->mutex );
-    OGR_L_ResetReading( hLayer );
-  }
-  ds->refCount ++;
-  return layer;
-}
-
-QgsOgrLayerUniquePtr QgsOgrLayer::CreateForSql(
-  const QgsOgrProviderUtils::DatasetIdentification &ident,
-  const QString &sql,
-  QgsOgrProviderUtils::DatasetWithLayers *ds,
-  OGRLayerH hLayer )
-{
-  QgsOgrLayerUniquePtr layer( new QgsOgrLayer() );
-  layer->ident = ident;
-  layer->isSqlLayer = true;
-  layer->sql = sql;
-  layer->ds = ds;
-  layer->hLayer = hLayer;
-  {
-    QMutexLocker locker( &ds->mutex );
-    OGR_L_ResetReading( hLayer );
-  }
-  ds->refCount ++;
-  return layer;
-}
-
-int QgsOgrLayer::GetLayerCount()
-{
-  QMutexLocker locker( &ds->mutex );
-  return GDALDatasetGetLayerCount( ds->hDS );
-}
-
-GDALDriverH QgsOgrLayer::driver()
-{
-  return GDALGetDatasetDriver( ds->hDS );
-}
-
-QString  QgsOgrLayer::driverName()
-{
-  return QString::fromUtf8( GDALGetDriverShortName( GDALGetDatasetDriver( ds->hDS ) ) );
-}
-
-QByteArray QgsOgrLayer::name()
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_GetName( hLayer );
-}
-
-void QgsOgrLayer::ResetReading()
-{
-  QMutexLocker locker( &ds->mutex );
-  OGR_L_ResetReading( hLayer );
-}
-
-QByteArray QgsOgrLayer::GetFIDColumn()
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_GetFIDColumn( hLayer );
-}
-
-OGRSpatialReferenceH QgsOgrLayer::GetSpatialRef()
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_GetSpatialRef( hLayer );
-}
-
-OGRFeatureH QgsOgrLayer::GetNextFeature()
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_GetNextFeature( hLayer );
-}
-
-OGRFeatureH QgsOgrLayer::GetFeature( GIntBig fid )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_GetFeature( hLayer, fid );
-}
-
-QgsOgrFeatureDefn &QgsOgrLayer::GetLayerDefn()
-{
-  return oFDefn;
-}
-
-GIntBig QgsOgrLayer::GetFeatureCount( bool force )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_GetFeatureCount( hLayer, force );
-}
-
-GIntBig QgsOgrLayer::GetApproxFeatureCount()
-{
-  QMutexLocker locker( &ds->mutex );
-
-  // OGR_L_GetFeatureCount() can be super slow on huge geopackage files
-  // so implement some approximation strategy that has reasonable runtime.
-  QString driverName = GDALGetDriverShortName( GDALGetDatasetDriver( ds->hDS ) );
-  if ( driverName == QLatin1String( "GPKG" ) )
-  {
-    CPLPushErrorHandler( CPLQuietErrorHandler );
-    OGRLayerH hSqlLayer = GDALDatasetExecuteSQL(
-                            ds->hDS, "SELECT 1 FROM gpkg_ogr_contents LIMIT 0", nullptr, nullptr );
-    CPLPopErrorHandler();
-    if ( hSqlLayer )
-    {
-      GDALDatasetReleaseResultSet( ds->hDS, hSqlLayer );
-      return OGR_L_GetFeatureCount( hLayer, TRUE );
-    }
-
-    // Enumerate features up to a limit of 100000.
-    const GIntBig nLimit = CPLAtoGIntBig(
-                             CPLGetConfigOption( "QGIS_GPKG_FC_THRESHOLD", "100000" ) );
-    QByteArray layerName = OGR_L_GetName( hLayer );
-    QByteArray sql( "SELECT COUNT(*) FROM (SELECT 1 FROM " );
-    sql += QgsOgrProviderUtils::quotedIdentifier( layerName, driverName );
-    sql += " LIMIT ";
-    sql += CPLSPrintf( CPL_FRMT_GIB, nLimit );
-    sql += ")";
-    hSqlLayer = GDALDatasetExecuteSQL( ds->hDS, sql, nullptr, nullptr );
-    GIntBig res = -1;
-    if ( hSqlLayer )
-    {
-      gdal::ogr_feature_unique_ptr fet( OGR_L_GetNextFeature( hSqlLayer ) );
-      if ( fet )
-      {
-        res = OGR_F_GetFieldAsInteger64( fet.get(), 0 );
-      }
-      GDALDatasetReleaseResultSet( ds->hDS, hSqlLayer );
-    }
-    if ( res >= 0 && res < nLimit )
-    {
-      // Less than 100000 features ? This is the final count
-      return res;
-    }
-    if ( res == nLimit )
-    {
-      // If we reach the threshold, then use the min and max values of the rowid
-      // hoping there are not a lot of holes.
-      // Do it in 2 separate SQL queries otherwise SQLite apparently does a
-      // full table scan...
-      sql = "SELECT MAX(ROWID) FROM ";
-      sql += QgsOgrProviderUtils::quotedIdentifier( layerName, driverName );
-      hSqlLayer = GDALDatasetExecuteSQL( ds->hDS, sql, nullptr, nullptr );
-      GIntBig maxrowid = -1;
-      if ( hSqlLayer )
-      {
-        gdal::ogr_feature_unique_ptr fet( OGR_L_GetNextFeature( hSqlLayer ) );
-        if ( fet )
-        {
-          maxrowid = OGR_F_GetFieldAsInteger64( fet.get(), 0 );
-        }
-        GDALDatasetReleaseResultSet( ds->hDS, hSqlLayer );
-      }
-
-      sql = "SELECT MIN(ROWID) FROM ";
-      sql += QgsOgrProviderUtils::quotedIdentifier( layerName, driverName );
-      hSqlLayer = GDALDatasetExecuteSQL( ds->hDS, sql, nullptr, nullptr );
-      GIntBig minrowid = 0;
-      if ( hSqlLayer )
-      {
-        gdal::ogr_feature_unique_ptr fet( OGR_L_GetNextFeature( hSqlLayer ) );
-        if ( fet )
-        {
-          minrowid = OGR_F_GetFieldAsInteger64( fet.get(), 0 );
-        }
-        GDALDatasetReleaseResultSet( ds->hDS, hSqlLayer );
-      }
-
-      if ( maxrowid >= minrowid )
-      {
-        return maxrowid - minrowid + 1;
-      }
-    }
-  }
-  if ( driverName == QLatin1String( "OAPIF" ) || driverName == QLatin1String( "WFS3" ) )
-  {
-    return -1;
-  }
-
-  return OGR_L_GetFeatureCount( hLayer, TRUE );
-}
-
-OGRErr QgsOgrLayer::GetExtent( OGREnvelope *psExtent, bool bForce )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_GetExtent( hLayer, psExtent, bForce );
-}
-
-OGRGeometryH QgsOgrLayer::GetSpatialFilter()
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_GetSpatialFilter( hLayer );
-}
-
-void QgsOgrLayer::SetSpatialFilter( OGRGeometryH hGeometry )
-{
-  QMutexLocker locker( &ds->mutex );
-  OGR_L_SetSpatialFilter( hLayer, hGeometry );
-}
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-GDALDatasetH QgsOgrLayer::getDatasetHandleAndMutex( QMutex *&mutex )
-#else
-GDALDatasetH QgsOgrLayer::getDatasetHandleAndMutex( QRecursiveMutex *&mutex )
-#endif
-{
-  mutex = &( ds->mutex );
-  return ds->hDS;
-}
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-OGRLayerH QgsOgrLayer::getHandleAndMutex( QMutex *&mutex )
-#else
-OGRLayerH QgsOgrLayer::getHandleAndMutex( QRecursiveMutex *&mutex )
-#endif
-{
-  mutex = &( ds->mutex );
-  return hLayer;
-}
-
-OGRErr QgsOgrLayer::CreateFeature( OGRFeatureH hFeature )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_CreateFeature( hLayer, hFeature );
-}
-
-OGRErr QgsOgrLayer::SetFeature( OGRFeatureH hFeature )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_SetFeature( hLayer, hFeature );
-}
-
-OGRErr QgsOgrLayer::DeleteFeature( GIntBig fid )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_DeleteFeature( hLayer, fid );
-}
-
-OGRErr QgsOgrLayer::CreateField( OGRFieldDefnH hFieldDefn, bool bStrict )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_CreateField( hLayer, hFieldDefn, bStrict );
-}
-
-OGRErr QgsOgrLayer::DeleteField( int iField )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_DeleteField( hLayer, iField );
-}
-
-OGRErr QgsOgrLayer::AlterFieldDefn( int iField, OGRFieldDefnH hNewFieldDefn, int flags )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_AlterFieldDefn( hLayer, iField, hNewFieldDefn, flags );
-}
-
-int QgsOgrLayer::TestCapability( const char *cap )
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_TestCapability( hLayer, cap );
-}
-
-OGRErr QgsOgrLayer::StartTransaction()
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_StartTransaction( hLayer );
-}
-
-OGRErr QgsOgrLayer::CommitTransaction()
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_CommitTransaction( hLayer );
-}
-
-OGRErr QgsOgrLayer::RollbackTransaction()
-{
-  QMutexLocker locker( &ds->mutex );
-  return OGR_L_RollbackTransaction( hLayer );
-}
-
-OGRErr QgsOgrLayer::SyncToDisk()
-{
-  QMutexLocker locker( &ds->mutex );
-
-  OGRErr eErr;
-#if GDAL_VERSION_NUM >= GDAL_COMPUTE_VERSION(3,1,0) && GDAL_VERSION_NUM <= GDAL_COMPUTE_VERSION(3,1,3)
-  // Workaround bug in GDAL 3.1.0 to 3.1.3 that creates XLSX and ODS files incompatible with LibreOffice due to use of ZIP64
-  QString drvName = GDALGetDriverShortName( GDALGetDatasetDriver( ds->hDS ) );
-  if ( drvName == QLatin1String( "XLSX" ) ||
-       drvName == QLatin1String( "ODS" ) )
-  {
-    CPLSetThreadLocalConfigOption( "CPL_CREATE_ZIP64", "NO" );
-    eErr = OGR_L_SyncToDisk( hLayer );
-    CPLSetThreadLocalConfigOption( "CPL_CREATE_ZIP64", nullptr );
-  }
-  else
-#endif
-  {
-    eErr = OGR_L_SyncToDisk( hLayer );
-  }
-
-  return eErr;
-}
-
-void QgsOgrLayer::ExecuteSQLNoReturn( const QByteArray &sql )
-{
-  QMutexLocker locker( &ds->mutex );
-  OGRLayerH hSqlLayer = GDALDatasetExecuteSQL( ds->hDS,
-                        sql.constData(),
-                        nullptr,
-                        nullptr );
-  GDALDatasetReleaseResultSet( ds->hDS, hSqlLayer );
-}
-
-QgsOgrLayerUniquePtr QgsOgrLayer::ExecuteSQL( const QByteArray &sql )
-{
-  QMutexLocker locker( &ds->mutex );
-  OGRLayerH hSqlLayer = GDALDatasetExecuteSQL( ds->hDS,
-                        sql.constData(),
-                        nullptr,
-                        nullptr );
-  if ( !hSqlLayer )
-    return nullptr;
-
-  return QgsOgrLayer::CreateForSql( ident,
-                                    QString::fromUtf8( sql ),
-                                    ds,
-                                    hSqlLayer );
-}
-
-QString QgsOgrLayer::GetMetadataItem( const QString &key, const QString &domain )
-{
-  QMutexLocker locker( &ds->mutex );
-  return GDALGetMetadataItem( hLayer, key.toUtf8().constData(),
-                              domain.toUtf8().constData() );
-}
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-QMutex &QgsOgrFeatureDefn::mutex()
-#else
-QRecursiveMutex &QgsOgrFeatureDefn::mutex()
-#endif
-{
-  return layer->mutex();
-}
-
-OGRFeatureDefnH QgsOgrFeatureDefn::get()
-{
-  if ( !hDefn )
-  {
-    QMutexLocker locker( &mutex() );
-    hDefn = OGR_L_GetLayerDefn( layer->hLayer );
-  }
-  return hDefn;
-}
-
-
-int QgsOgrFeatureDefn::GetFieldCount()
-{
-  QMutexLocker locker( &mutex() );
-  return OGR_FD_GetFieldCount( get() );
-}
-
-OGRFieldDefnH QgsOgrFeatureDefn::GetFieldDefn( int idx )
-{
-  QMutexLocker locker( &mutex() );
-  return OGR_FD_GetFieldDefn( get(), idx );
-}
-
-int QgsOgrFeatureDefn::GetFieldIndex( const QByteArray &name )
-{
-  QMutexLocker locker( &mutex() );
-  return OGR_FD_GetFieldIndex( get(), name.constData() );
-}
-
-int QgsOgrFeatureDefn::GetGeomFieldCount()
-{
-  QMutexLocker locker( &mutex() );
-  return OGR_FD_GetGeomFieldCount( get() );
-}
-
-OGRGeomFieldDefnH  QgsOgrFeatureDefn::GetGeomFieldDefn( int idx )
-{
-  QMutexLocker locker( &mutex() );
-  return OGR_FD_GetGeomFieldDefn( get(), idx );
-}
-
-OGRwkbGeometryType QgsOgrFeatureDefn::GetGeomType()
-{
-  QMutexLocker locker( &mutex() );
-  return OGR_FD_GetGeomType( get() );
-}
-
-OGRFeatureH QgsOgrFeatureDefn::CreateFeature()
-{
-  QMutexLocker locker( &mutex() );
-  return OGR_F_Create( get() );
-}
-
-// ---------------------------------------------------------------------------
-
-static
-QgsOgrLayerUniquePtr LoadDataSourceAndLayer( const QString &uri,
-    QString &errCause )
-{
-  bool isSubLayer;
-  int layerIndex;
-  QString layerName;
-  QString subsetString;
-  OGRwkbGeometryType ogrGeometryType;
-  QStringList openOptions;
-  QString filePath = AnalyzeURI( uri,
-                                 isSubLayer,
-                                 layerIndex,
-                                 layerName,
-                                 subsetString,
-                                 ogrGeometryType,
-                                 openOptions );
-
-  if ( !layerName.isEmpty() )
-  {
-    return QgsOgrProviderUtils::getLayer( filePath, true, QStringList(), layerName, errCause, true );
-  }
-  else
-  {
-    return QgsOgrProviderUtils::getLayer( filePath, true, QStringList(), layerIndex, errCause, true );
-  }
-}
-
-
-bool QgsOgrProviderMetadata::saveStyle(
-  const QString &uri, const QString &qmlStyle, const QString &sldStyle,
-  const QString &styleName, const QString &styleDescription,
-  const QString &uiFileContent, bool useAsDefault, QString &errCause )
-{
-  QgsOgrLayerUniquePtr userLayer = LoadDataSourceAndLayer( uri, errCause );
-  if ( !userLayer )
-    return false;
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-  QMutex *mutex = nullptr;
-#else
-  QRecursiveMutex *mutex = nullptr;
-#endif
-  OGRLayerH hUserLayer = userLayer->getHandleAndMutex( mutex );
-  GDALDatasetH hDS = userLayer->getDatasetHandleAndMutex( mutex );
-  QMutexLocker locker( mutex );
-
-  // check if layer_styles table already exist
-  OGRLayerH hLayer = GDALDatasetGetLayerByName( hDS, "layer_styles" );
-  if ( !hLayer )
-  {
-    // if not create it
-    // Note: we use the same schema as in the SpatiaLite and postgres providers
-    //for cross interoperability
-
-    char **options = nullptr;
-    // TODO: might need change if other drivers than GPKG / SQLite
-    options = CSLSetNameValue( options, "FID", "id" );
-    hLayer = GDALDatasetCreateLayer( hDS, "layer_styles", nullptr, wkbNone, options );
-    QgsOgrProviderUtils::invalidateCachedDatasets( QString::fromUtf8( GDALGetDescription( hDS ) ) );
-    CSLDestroy( options );
-    if ( !hLayer )
-    {
-      errCause = QObject::tr( "Unable to save layer style. It's not possible to create the destination table on the database." );
-      return false;
-    }
-    bool ok = true;
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "f_table_catalog", OFTString ) );
-      OGR_Fld_SetWidth( fld.get(), 256 );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "f_table_schema", OFTString ) );
-      OGR_Fld_SetWidth( fld.get(), 256 );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "f_table_name", OFTString ) );
-      OGR_Fld_SetWidth( fld.get(), 256 );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "f_geometry_column", OFTString ) );
-      OGR_Fld_SetWidth( fld.get(), 256 );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "styleName", OFTString ) );
-      OGR_Fld_SetWidth( fld.get(), 30 );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "styleQML", OFTString ) );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "styleSLD", OFTString ) );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "useAsDefault", OFTInteger ) );
-      OGR_Fld_SetSubType( fld.get(), OFSTBoolean );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "description", OFTString ) );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "owner", OFTString ) );
-      OGR_Fld_SetWidth( fld.get(), 30 );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "ui", OFTString ) );
-      OGR_Fld_SetWidth( fld.get(), 30 );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    {
-      gdal::ogr_field_def_unique_ptr fld( OGR_Fld_Create( "update_time", OFTDateTime ) );
-      OGR_Fld_SetDefault( fld.get(), "CURRENT_TIMESTAMP" );
-      ok &= OGR_L_CreateField( hLayer, fld.get(), true ) == OGRERR_NONE;
-    }
-    if ( !ok )
-    {
-      errCause = QObject::tr( "Unable to save layer style. It's not possible to create the destination table on the database." );
-      return false;
-    }
-  }
-
-  QString realStyleName =
-    styleName.isEmpty() ? QString( OGR_L_GetName( hUserLayer ) ) : styleName;
-
-  OGRFeatureDefnH hLayerDefn = OGR_L_GetLayerDefn( hLayer );
-
-  if ( useAsDefault )
-  {
-    QString oldDefaultQuery = QStringLiteral( "useAsDefault = 1 AND f_table_schema=''"
-                              " AND f_table_name=%1"
-                              " AND f_geometry_column=%2" )
-                              .arg( QgsOgrProviderUtils::quotedValue( QString( OGR_L_GetName( hUserLayer ) ) ) )
-                              .arg( QgsOgrProviderUtils::quotedValue( QString( OGR_L_GetGeometryColumn( hUserLayer ) ) ) );
-    OGR_L_SetAttributeFilter( hLayer, oldDefaultQuery.toUtf8().constData() );
-    gdal::ogr_feature_unique_ptr hFeature( OGR_L_GetNextFeature( hLayer ) );
-    if ( hFeature )
-    {
-      OGR_F_SetFieldInteger( hFeature.get(),
-                             OGR_FD_GetFieldIndex( hLayerDefn, "useAsDefault" ),
-                             0 );
-      bool ok = OGR_L_SetFeature( hLayer, hFeature.get() ) == 0;
-      if ( !ok )
-      {
-        QgsDebugMsg( QStringLiteral( "Could not unset previous useAsDefault style" ) );
-      }
-    }
-  }
-
-  QString checkQuery = QStringLiteral( "f_table_schema=''"
-                                       " AND f_table_name=%1"
-                                       " AND f_geometry_column=%2"
-                                       " AND styleName=%3" )
-                       .arg( QgsOgrProviderUtils::quotedValue( QString( OGR_L_GetName( hUserLayer ) ) ) )
-                       .arg( QgsOgrProviderUtils::quotedValue( QString( OGR_L_GetGeometryColumn( hUserLayer ) ) ) )
-                       .arg( QgsOgrProviderUtils::quotedValue( realStyleName ) );
-  OGR_L_SetAttributeFilter( hLayer, checkQuery.toUtf8().constData() );
-  OGR_L_ResetReading( hLayer );
-  gdal::ogr_feature_unique_ptr hFeature( OGR_L_GetNextFeature( hLayer ) );
-  bool bNew = true;
-
-  if ( hFeature )
-  {
-    QgsSettings settings;
-    // Only used in tests. Do not define it for interactive implication
-    QVariant overwriteStyle = settings.value( QStringLiteral( "qgis/overwriteStyle" ) );
-    if ( ( !overwriteStyle.isNull() && !overwriteStyle.toBool() ) ||
-         ( overwriteStyle.isNull() &&
-           QMessageBox::question( nullptr, QObject::tr( "Save style in database" ),
-                                  QObject::tr( "A style named \"%1\" already exists in the database for this layer. Do you want to overwrite it?" )
-                                  .arg( realStyleName ),
-                                  QMessageBox::Yes | QMessageBox::No ) == QMessageBox::No ) )
-    {
-      errCause = QObject::tr( "Operation aborted" );
-      return false;
-    }
-    bNew = false;
-  }
-  else
-  {
-    hFeature.reset( OGR_F_Create( hLayerDefn ) );
-    OGR_F_SetFieldString( hFeature.get(),
-                          OGR_FD_GetFieldIndex( hLayerDefn, "f_table_catalog" ),
-                          "" );
-    OGR_F_SetFieldString( hFeature.get(),
-                          OGR_FD_GetFieldIndex( hLayerDefn, "f_table_schema" ),
-                          "" );
-    OGR_F_SetFieldString( hFeature.get(),
-                          OGR_FD_GetFieldIndex( hLayerDefn, "f_table_name" ),
-                          OGR_L_GetName( hUserLayer ) );
-    OGR_F_SetFieldString( hFeature.get(),
-                          OGR_FD_GetFieldIndex( hLayerDefn, "f_geometry_column" ),
-                          OGR_L_GetGeometryColumn( hUserLayer ) );
-    OGR_F_SetFieldString( hFeature.get(),
-                          OGR_FD_GetFieldIndex( hLayerDefn, "styleName" ),
-                          realStyleName.toUtf8().constData() );
-    if ( !uiFileContent.isEmpty() )
-    {
-      OGR_F_SetFieldString( hFeature.get(),
-                            OGR_FD_GetFieldIndex( hLayerDefn, "ui" ),
-                            uiFileContent.toUtf8().constData() );
-    }
-  }
-  OGR_F_SetFieldString( hFeature.get(),
-                        OGR_FD_GetFieldIndex( hLayerDefn, "styleQML" ),
-                        qmlStyle.toUtf8().constData() );
-  OGR_F_SetFieldString( hFeature.get(),
-                        OGR_FD_GetFieldIndex( hLayerDefn, "styleSLD" ),
-                        sldStyle.toUtf8().constData() );
-  OGR_F_SetFieldInteger( hFeature.get(),
-                         OGR_FD_GetFieldIndex( hLayerDefn, "useAsDefault" ),
-                         useAsDefault ? 1 : 0 );
-  OGR_F_SetFieldString( hFeature.get(),
-                        OGR_FD_GetFieldIndex( hLayerDefn, "description" ),
-                        ( styleDescription.isEmpty() ? QDateTime::currentDateTime().toString() : styleDescription ).toUtf8().constData() );
-  OGR_F_SetFieldString( hFeature.get(),
-                        OGR_FD_GetFieldIndex( hLayerDefn, "owner" ),
-                        "" );
-
-  bool bFeatureOK;
-  if ( bNew )
-    bFeatureOK = OGR_L_CreateFeature( hLayer, hFeature.get() ) == OGRERR_NONE;
-  else
-    bFeatureOK = OGR_L_SetFeature( hLayer, hFeature.get() ) == OGRERR_NONE;
-
-  if ( !bFeatureOK )
-  {
-    QgsMessageLog::logMessage( QObject::tr( "Error updating style" ) );
-    errCause = QObject::tr( "Error looking for style. The query was logged" );
-    return false;
-  }
-
-  return true;
-}
-
-
-bool QgsOgrProviderMetadata::deleteStyleById( const QString &uri, QString styleId, QString &errCause )
-{
-  QgsDataSourceUri dsUri( uri );
-  bool deleted;
-
-  QgsOgrLayerUniquePtr userLayer = LoadDataSourceAndLayer( uri, errCause );
-  if ( !userLayer )
-    return false;
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-  QMutex *mutex = nullptr;
-#else
-  QRecursiveMutex *mutex = nullptr;
-#endif
-  GDALDatasetH hDS = userLayer->getDatasetHandleAndMutex( mutex );
-  QMutexLocker locker( mutex );
-
-  // check if layer_styles table already exist
-  OGRLayerH hLayer = GDALDatasetGetLayerByName( hDS, "layer_styles" );
-  if ( !hLayer )
-  {
-    errCause = QObject::tr( "Connection to database failed: %1" ).arg( dsUri.uri() );
-    deleted = false;
-  }
-  else
-  {
-    if ( OGR_L_DeleteFeature( hLayer, styleId.toInt() ) != OGRERR_NONE )
-    {
-      errCause = QObject::tr( "Error executing the delete query." );
-      deleted = false;
-    }
-    else
-    {
-      deleted = true;
-    }
-  }
-  return deleted;
-}
-
-static
-bool LoadDataSourceLayerStylesAndLayer( const QString &uri,
-                                        QgsOgrLayerUniquePtr &layerStyles,
-                                        QgsOgrLayerUniquePtr &userLayer,
-                                        QString &errCause )
-{
-  bool isSubLayer;
-  int layerIndex;
-  QString layerName;
-  QString subsetString;
-  OGRwkbGeometryType ogrGeometryType;
-  QStringList openOptions;
-  QString filePath = AnalyzeURI( uri,
-                                 isSubLayer,
-                                 layerIndex,
-                                 layerName,
-                                 subsetString,
-                                 ogrGeometryType,
-                                 openOptions );
-
-  layerStyles =
-    QgsOgrProviderUtils::getLayer( filePath, "layer_styles", errCause );
-  userLayer = nullptr;
-  if ( !layerStyles )
-  {
-    errCause = QObject::tr( "Cannot find layer_styles layer" );
-    return false;
-  }
-
-  if ( !layerName.isEmpty() )
-  {
-    userLayer = QgsOgrProviderUtils::getLayer( filePath, layerName, errCause );
-  }
-  else
-  {
-    userLayer = QgsOgrProviderUtils::getLayer( filePath, layerIndex, errCause );
-  }
-  if ( !userLayer )
-  {
-    layerStyles.reset();
-    return false;
-  }
-  return true;
-}
-
-
-QString QgsOgrProviderMetadata::loadStyle( const QString &uri, QString &errCause )
-{
-  QgsOgrLayerUniquePtr layerStyles;
-  QgsOgrLayerUniquePtr userLayer;
-  if ( !LoadDataSourceLayerStylesAndLayer( uri, layerStyles, userLayer, errCause ) )
-  {
-    return QString();
-  }
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-  QMutex *mutex1 = nullptr;
-  QMutex *mutex2 = nullptr;
-#else
-  QRecursiveMutex *mutex1 = nullptr;
-  QRecursiveMutex *mutex2 = nullptr;
-#endif
-  OGRLayerH hLayer = layerStyles->getHandleAndMutex( mutex1 );
-  OGRLayerH hUserLayer = userLayer->getHandleAndMutex( mutex2 );
-  QMutexLocker lock1( mutex1 );
-  QMutexLocker lock2( mutex2 );
-
-  QString selectQmlQuery = QStringLiteral( "f_table_schema=''"
-                           " AND f_table_name=%1"
-                           " AND f_geometry_column=%2"
-                           " ORDER BY CASE WHEN useAsDefault THEN 1 ELSE 2 END"
-                           ",update_time DESC LIMIT 1" )
-                           .arg( QgsOgrProviderUtils::quotedValue( QString( OGR_L_GetName( hUserLayer ) ) ) )
-                           .arg( QgsOgrProviderUtils::quotedValue( QString( OGR_L_GetGeometryColumn( hUserLayer ) ) ) );
-  OGR_L_SetAttributeFilter( hLayer, selectQmlQuery.toUtf8().constData() );
-  OGR_L_ResetReading( hLayer );
-  OGRFeatureDefnH hLayerDefn = OGR_L_GetLayerDefn( hLayer );
-  QString styleQML;
-  qlonglong moreRecentTimestamp = 0;
-  while ( true )
-  {
-    gdal::ogr_feature_unique_ptr hFeat( OGR_L_GetNextFeature( hLayer ) );
-    if ( !hFeat )
-      break;
-    if ( OGR_F_GetFieldAsInteger( hFeat.get(), OGR_FD_GetFieldIndex( hLayerDefn, "useAsDefault" ) ) )
-    {
-      styleQML = QString::fromUtf8(
-                   OGR_F_GetFieldAsString( hFeat.get(), OGR_FD_GetFieldIndex( hLayerDefn, "styleQML" ) ) );
-      break;
-    }
-
-    int  year, month, day, hour, minute, second, TZ;
-    OGR_F_GetFieldAsDateTime( hFeat.get(), OGR_FD_GetFieldIndex( hLayerDefn, "update_time" ),
-                              &year, &month, &day, &hour, &minute, &second, &TZ );
-    qlonglong ts = second + minute * 60 + hour * 3600 + day * 24 * 3600 +
-                   static_cast<qlonglong>( month ) * 31 * 24 * 3600 + static_cast<qlonglong>( year ) * 12 * 31 * 24 * 3600;
-    if ( ts > moreRecentTimestamp )
-    {
-      moreRecentTimestamp = ts;
-      styleQML = QString::fromUtf8(
-                   OGR_F_GetFieldAsString( hFeat.get(), OGR_FD_GetFieldIndex( hLayerDefn, "styleQML" ) ) );
-
-    }
-  }
-
-  return styleQML;
-}
-
-int QgsOgrProviderMetadata::listStyles(
-  const QString &uri, QStringList &ids, QStringList &names,
-  QStringList &descriptions, QString &errCause )
-{
-  bool isSubLayer;
-  int layerIndex;
-  QString layerName;
-  QString subsetString;
-  OGRwkbGeometryType ogrGeometryType;
-  QStringList openOptions;
-  QString filePath = AnalyzeURI( uri,
-                                 isSubLayer,
-                                 layerIndex,
-                                 layerName,
-                                 subsetString,
-                                 ogrGeometryType,
-                                 openOptions );
-
-  QgsOgrLayerUniquePtr userLayer;
-  if ( !layerName.isEmpty() )
-  {
-    userLayer = QgsOgrProviderUtils::getLayer( filePath, layerName, errCause );
-  }
-  else
-  {
-    userLayer = QgsOgrProviderUtils::getLayer( filePath, layerIndex, errCause );
-  }
-  if ( !userLayer )
-  {
-    return -1;
-  }
-
-  QgsOgrLayerUniquePtr layerStyles =
-    QgsOgrProviderUtils::getLayer( filePath, "layer_styles", errCause );
-  if ( !layerStyles )
-  {
-    QgsMessageLog::logMessage( QObject::tr( "No styles available on DB" ) );
-    errCause = QObject::tr( "No styles available on DB" );
-    return 0;
-  }
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-  QMutex *mutex1 = nullptr;
-  QMutex *mutex2 = nullptr;
-#else
-  QRecursiveMutex *mutex1 = nullptr;
-  QRecursiveMutex *mutex2 = nullptr;
-#endif
-
-  OGRLayerH hLayer = layerStyles->getHandleAndMutex( mutex1 );
-  QMutexLocker lock1( mutex1 );
-  OGRLayerH hUserLayer = userLayer->getHandleAndMutex( mutex2 );
-  QMutexLocker lock2( mutex2 );
-
-  if ( OGR_L_GetFeatureCount( hLayer, TRUE ) == 0 )
-  {
-    QgsMessageLog::logMessage( QObject::tr( "No styles available on DB" ) );
-    errCause = QObject::tr( "No styles available on DB" );
-    return 0;
-  }
-
-  OGRFeatureDefnH hLayerDefn = OGR_L_GetLayerDefn( hLayer );
-
-  OGR_L_ResetReading( hLayer );
-
-  QList<qlonglong> listTimestamp;
-  QMap<int, QString> mapIdToStyleName;
-  QMap<int, QString> mapIdToDescription;
-  QMap<qlonglong, QList<int> > mapTimestampToId;
-  int numberOfRelatedStyles = 0;
-  while ( true )
-  {
-    gdal::ogr_feature_unique_ptr hFeature( OGR_L_GetNextFeature( hLayer ) );
-    if ( !hFeature )
-      break;
-
-    QString tableName( QString::fromUtf8(
-                         OGR_F_GetFieldAsString( hFeature.get(),
-                             OGR_FD_GetFieldIndex( hLayerDefn, "f_table_name" ) ) ) );
-    QString geometryColumn( QString::fromUtf8(
-                              OGR_F_GetFieldAsString( hFeature.get(),
-                                  OGR_FD_GetFieldIndex( hLayerDefn, "f_geometry_column" ) ) ) );
-    QString styleName( QString::fromUtf8(
-                         OGR_F_GetFieldAsString( hFeature.get(),
-                             OGR_FD_GetFieldIndex( hLayerDefn, "styleName" ) ) ) );
-    QString description( QString::fromUtf8(
-                           OGR_F_GetFieldAsString( hFeature.get(),
-                               OGR_FD_GetFieldIndex( hLayerDefn, "description" ) ) ) );
-    int fid = static_cast<int>( OGR_F_GetFID( hFeature.get() ) );
-    if ( tableName == QString::fromUtf8( OGR_L_GetName( hUserLayer ) ) &&
-         geometryColumn == QString::fromUtf8( OGR_L_GetGeometryColumn( hUserLayer ) ) )
-    {
-      // Append first all related styles
-      QString id( QString::number( fid ) );
-      ids.append( id );
-      names.append( styleName );
-      descriptions.append( description );
-      ++ numberOfRelatedStyles;
-    }
-    else
-    {
-      int  year, month, day, hour, minute, second, TZ;
-      OGR_F_GetFieldAsDateTime( hFeature.get(), OGR_FD_GetFieldIndex( hLayerDefn, "update_time" ),
-                                &year, &month, &day, &hour, &minute, &second, &TZ );
-      qlonglong ts = second + minute * 60 + hour * 3600 + day * 24 * 3600 +
-                     static_cast<qlonglong>( month ) * 31 * 24 * 3600 + static_cast<qlonglong>( year ) * 12 * 31 * 24 * 3600;
-
-      listTimestamp.append( ts );
-      mapIdToStyleName[fid] = styleName;
-      mapIdToDescription[fid] = description;
-      mapTimestampToId[ts].append( fid );
-    }
-  }
-
-  std::sort( listTimestamp.begin(), listTimestamp.end() );
-  // Sort from most recent to least recent
-  for ( int i = listTimestamp.size() - 1; i >= 0; i-- )
-  {
-    const QList<int> &listId = mapTimestampToId[listTimestamp[i]];
-    for ( int j = 0; j < listId.size(); j++ )
-    {
-      int fid = listId[j];
-      QString id( QString::number( fid ) );
-      ids.append( id );
-      names.append( mapIdToStyleName[fid] );
-      descriptions.append( mapIdToDescription[fid] );
-    }
-  }
-
-  return numberOfRelatedStyles;
-}
-
-QString QgsOgrProviderMetadata::getStyleById( const QString &uri, QString styleId, QString &errCause )
-{
-  QgsOgrLayerUniquePtr layerStyles;
-  QgsOgrLayerUniquePtr userLayer;
-  if ( !LoadDataSourceLayerStylesAndLayer( uri, layerStyles, userLayer, errCause ) )
-  {
-    return QString();
-  }
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-  QMutex *mutex1 = nullptr;
-#else
-  QRecursiveMutex *mutex1 = nullptr;
-#endif
-
-  OGRLayerH hLayer = layerStyles->getHandleAndMutex( mutex1 );
-  QMutexLocker lock1( mutex1 );
-
-  bool ok;
-  int id = styleId.toInt( &ok );
-  if ( !ok )
-  {
-    errCause = QObject::tr( "Invalid style identifier" );
-    return QString();
-  }
-
-  gdal::ogr_feature_unique_ptr hFeature( OGR_L_GetFeature( hLayer, id ) );
-  if ( !hFeature )
-  {
-    errCause = QObject::tr( "No style corresponding to style identifier" );
-    return QString();
-  }
-
-  OGRFeatureDefnH hLayerDefn = OGR_L_GetLayerDefn( hLayer );
-  QString styleQML( QString::fromUtf8(
-                      OGR_F_GetFieldAsString( hFeature.get(),
-                          OGR_FD_GetFieldIndex( hLayerDefn, "styleQML" ) ) ) );
-
-  return styleQML;
-}
-
-bool QgsOgrProviderMetadata::saveLayerMetadata( const QString &uri, const QgsLayerMetadata &metadata, QString &errorMessage )
-{
-  const QVariantMap parts = decodeUri( uri );
-  const QString path = parts.value( QStringLiteral( "path" ) ).toString();
-  if ( !path.isEmpty() && QFileInfo::exists( path ) )
-  {
-    // export metadata to XML
-    QDomImplementation domImplementation;
-    QDomDocumentType documentType = domImplementation.createDocumentType( QStringLiteral( "qgis" ), QStringLiteral( "http://mrcc.com/qgis.dtd" ), QStringLiteral( "SYSTEM" ) );
-    QDomDocument document( documentType );
-
-    QDomElement rootNode = document.createElement( QStringLiteral( "qgis" ) );
-    rootNode.setAttribute( QStringLiteral( "version" ), Qgis::version() );
-    document.appendChild( rootNode );
-
-    if ( !metadata.writeMetadataXml( rootNode, document ) )
-    {
-      errorMessage = QObject::tr( "Error exporting metadata to XML" );
-      return false;
-    }
-
-    QString metadataXml;
-    QTextStream textStream( &metadataXml );
-    document.save( textStream, 2 );
-
-    QFileInfo fi( path );
-    if ( fi.suffix().compare( QLatin1String( "gpkg" ), Qt::CaseInsensitive ) == 0 )
-    {
-      const QString layerName = parts.value( QStringLiteral( "layerName" ) ).toString();
-      QgsOgrLayerUniquePtr userLayer;
-      userLayer = QgsOgrProviderUtils::getLayer( path, true, QStringList(), layerName, errorMessage, true );
-      if ( !userLayer )
-        return false;
-
-#if QT_VERSION < QT_VERSION_CHECK(5, 14, 0)
-      QMutex *mutex = nullptr;
-#else
-      QRecursiveMutex *mutex = nullptr;
-#endif
-      // Returns native OGRLayerH object with the mutex to lock when using it
-      OGRLayerH hLayer = userLayer->getHandleAndMutex( mutex );
-      QMutexLocker locker( mutex );
-
-      // These are special keys which get stored into the gpkg_contents table
-      if ( !metadata.abstract().isEmpty() )
-        GDALSetMetadataItem( hLayer, "DESCRIPTION", metadata.abstract().toUtf8().constData(), nullptr );
-      if ( !metadata.identifier().isEmpty() )
-        GDALSetMetadataItem( hLayer, "IDENTIFIER", metadata.identifier().toUtf8().constData(), nullptr );
-
-      // we write a simple piece of GDAL metadata too -- this is solely to delegate responsibility of
-      // creating all the metadata tables to GDAL! We will remove it once done.
-      if ( GDALSetMetadataItem( hLayer, "QGIS_VERSION", Qgis::version().toLocal8Bit().constData(), nullptr ) == CE_None )
-      {
-        // so far so good, ready to throw the whole of the QGIS layer XML into the metadata table!
-
-        // first we need to check if there's already a corresponding entry in gpkg_metadata -- if so, we need to update it.
-        QString sql = QStringLiteral( "SELECT id from gpkg_metadata LEFT JOIN gpkg_metadata_reference ON "
-                                      "(gpkg_metadata_reference.table_name = %1 AND gpkg_metadata.id = gpkg_metadata_reference.md_file_id) "
-                                      "WHERE md_standard_uri = %2 and reference_scope = %3" ).arg(
-                        QgsSqliteUtils::quotedString( layerName ),
-                        QgsSqliteUtils::quotedString( QStringLiteral( "http://mrcc.com/qgis.dtd" ) ),
-                        QgsSqliteUtils::quotedString( QStringLiteral( "table" ) ) );
-        int existingRowId = -1;
-        if ( QgsOgrLayerUniquePtr l = userLayer->ExecuteSQL( sql.toLocal8Bit().constData() ) )
-        {
-          // retrieve inserted row id
-          gdal::ogr_feature_unique_ptr f( l->GetNextFeature() );
-          if ( f )
-          {
-            bool ok = false;
-            QVariant res = QgsOgrUtils::getOgrFeatureAttribute( f.get(), QgsField( QString(), QVariant::String ), 0, nullptr, &ok );
-            if ( ok )
-            {
-              existingRowId = res.toInt( &ok );
-              if ( !ok )
-                existingRowId = -1;
-            }
-          }
-        }
-
-        if ( existingRowId >= 0 )
-        {
-          // update existing row
-          sql = QStringLiteral( "UPDATE gpkg_metadata SET metadata=%1 where id=%2;" ).arg(
-                  QgsSqliteUtils::quotedString( metadataXml ) ).arg( existingRowId );
-          userLayer->ExecuteSQLNoReturn( sql.toLocal8Bit().constData() );
-          if ( CPLGetLastErrorType() != CE_None )
-          {
-            errorMessage = QStringLiteral( "%1 (%2): %3" ).arg( CPLGetLastErrorType() ).arg( CPLGetLastErrorNo() ).arg( CPLGetLastErrorMsg() );
-            return false;
-          }
-          else
-          {
-            // Remove QGIS_VERSION now that we are done
-            GDALSetMetadataItem( hLayer, "QGIS_VERSION", nullptr, nullptr );
-            return true;
-          }
-        }
-        else
-        {
-          // insert new details in metadata tables
-          sql = QStringLiteral( "INSERT INTO gpkg_metadata (md_scope, md_standard_uri, mime_type, metadata) VALUES (%1,%2,%3,%4);" )
-                .arg( QgsSqliteUtils::quotedString( QStringLiteral( "dataset" ) ),
-                      QgsSqliteUtils::quotedString( QStringLiteral( "http://mrcc.com/qgis.dtd" ) ),
-                      QgsSqliteUtils::quotedString( QStringLiteral( "text/xml" ) ),
-                      QgsSqliteUtils::quotedString( metadataXml ) );
-          userLayer->ExecuteSQLNoReturn( sql.toLocal8Bit().constData() );
-
-          sql = QStringLiteral( "SELECT last_insert_rowid();" );
-          int lastRowId = -1;
-          if ( QgsOgrLayerUniquePtr  l = userLayer->ExecuteSQL( sql.toLocal8Bit().constData() ) )
-          {
-            // retrieve inserted row id
-            gdal::ogr_feature_unique_ptr f( l->GetNextFeature() );
-            if ( f )
-            {
-              bool ok = false;
-              QVariant res = QgsOgrUtils::getOgrFeatureAttribute( f.get(), QgsField( QString(), QVariant::String ), 0, nullptr, &ok );
-              if ( !ok )
-              {
-                return false;
-              }
-              lastRowId = res.toInt();
-
-              sql = QStringLiteral( "INSERT INTO gpkg_metadata_reference (reference_scope, table_name, md_file_id) VALUES (%1,%2,%3);" )
-                    .arg( QgsSqliteUtils::quotedString( QStringLiteral( "table" ) ),
-                          QgsSqliteUtils::quotedString( layerName ) )
-                    .arg( lastRowId );
-              userLayer->ExecuteSQLNoReturn( sql.toLocal8Bit().constData() );
-
-              // Remove QGIS_VERSION now that we are done
-              GDALSetMetadataItem( hLayer, "QGIS_VERSION", nullptr, nullptr );
-              return true;
-            }
-          }
-          errorMessage = QStringLiteral( "Could not retrieve gpkg_metadata row id" );
-          return false;
-        }
-      }
-      else
-      {
-        errorMessage = QStringLiteral( "%1 (%2): %3" ).arg( CPLGetLastErrorType() ).arg( CPLGetLastErrorNo() ).arg( CPLGetLastErrorMsg() );
-        return false;
-      }
-    }
-    else
-    {
-      // file based, but not a geopackage -- store as .qmd sidecar file instead
-      // (possibly there's other formats outside of GPKG which also has some native means of storing metadata,
-      // which could be added for those formats before we resort to the sidecar approach!)
-      const QString qmdFileName = fi.dir().filePath( fi.completeBaseName() + QStringLiteral( ".qmd" ) );
-      QFile qmdFile( qmdFileName );
-      if ( qmdFile.open( QFile::WriteOnly | QFile::Truncate ) )
-      {
-        QTextStream fileStream( &qmdFile );
-        fileStream << metadataXml;
-        qmdFile.close();
-        return true;
-      }
-      else
-      {
-        errorMessage = tr( "ERROR: Failed to created default metadata file as %1. Check file permissions and retry." ).arg( qmdFileName );
-        return false;
-      }
-    }
-  }
-
-  throw QgsNotSupportedException( QObject::tr( "Storing metadata for the specified uri is not supported" ) );
-}
-
-// ---------------------------------------------------------------------------
-
-bool QgsOgrProviderUtils::deleteLayer( const QString &uri, QString &errCause )
-{
-  bool isSubLayer;
-  int layerIndex;
-  QString layerName;
-  QString subsetString;
-  OGRwkbGeometryType ogrGeometryType;
-  QStringList openOptions;
-  QString filePath = AnalyzeURI( uri,
-                                 isSubLayer,
-                                 layerIndex,
-                                 layerName,
-                                 subsetString,
-                                 ogrGeometryType,
-                                 openOptions );
-
-
-  GDALDatasetH hDS = GDALOpenEx( filePath.toUtf8().constData(), GDAL_OF_RASTER | GDAL_OF_VECTOR | GDAL_OF_UPDATE, nullptr, nullptr, nullptr );
-  if ( hDS  && ( ! layerName.isEmpty() || layerIndex != -1 ) )
-  {
-    // If we have got a name we convert it into an index
-    if ( ! layerName.isEmpty() )
-    {
-      layerIndex = -1;
-      for ( int i = 0; i < GDALDatasetGetLayerCount( hDS ); i++ )
-      {
-        OGRLayerH hL = GDALDatasetGetLayer( hDS, i );
-        if ( layerName == QString::fromUtf8( OGR_L_GetName( hL ) ) )
-        {
-          layerIndex = i;
-          break;
-        }
-      }
-    }
-    // Do delete!
-    if ( layerIndex != -1 )
-    {
-      OGRErr error = GDALDatasetDeleteLayer( hDS, layerIndex );
-      switch ( error )
-      {
-        case OGRERR_NOT_ENOUGH_DATA:
-          errCause = QObject::tr( "Not enough data to deserialize" );
-          break;
-        case OGRERR_NOT_ENOUGH_MEMORY:
-          errCause = QObject::tr( "Not enough memory" );
-          break;
-        case OGRERR_UNSUPPORTED_GEOMETRY_TYPE:
-          errCause = QObject::tr( "Unsupported geometry type" );
-          break;
-        case OGRERR_UNSUPPORTED_OPERATION:
-          errCause = QObject::tr( "Unsupported operation" );
-          break;
-        case OGRERR_CORRUPT_DATA:
-          errCause = QObject::tr( "Corrupt data" );
-          break;
-        case OGRERR_FAILURE:
-          errCause = QObject::tr( "Failure" );
-          break;
-        case OGRERR_UNSUPPORTED_SRS:
-          errCause = QObject::tr( "Unsupported SRS" );
-          break;
-        case OGRERR_INVALID_HANDLE:
-          errCause = QObject::tr( "Invalid handle" );
-          break;
-        case OGRERR_NON_EXISTING_FEATURE:
-          errCause = QObject::tr( "Non existing feature" );
-          break;
-        default:
-        case OGRERR_NONE:
-          errCause = QObject::tr( "Success" );
-          break;
-      }
-      errCause = QObject::tr( "GDAL result code: %1" ).arg( errCause );
-      return error == OGRERR_NONE;
-    }
-  }
-  // This should never happen:
-  errCause = QObject::tr( "Layer not found: %1" ).arg( uri );
-  return false;
-}
-
-void QgsOgrLayerReleaser::operator()( QgsOgrLayer *layer )
-{
-  QgsOgrProviderUtils::release( layer );
-}
-
-QgsTransaction *QgsOgrProviderMetadata::createTransaction( const QString &connString )
-{
-  auto ds = QgsOgrProviderUtils::getAlreadyOpenedDataset( connString );
-  if ( !ds )
-  {
-    QgsMessageLog::logMessage( QObject::tr( "Cannot open transaction on %1, since it is is not currently opened" ).arg( connString ),
-                               QObject::tr( "OGR" ), Qgis::MessageLevel::Critical );
-    return nullptr;
-  }
-
-  return new QgsOgrTransaction( connString, ds );
-}
-
-QgsGeoPackageProjectStorage *gGeoPackageProjectStorage = nullptr;   // when not null it is owned by QgsApplication::projectStorageRegistry()
-
-void QgsOgrProviderMetadata::initProvider()
-{
-  Q_ASSERT( !gGeoPackageProjectStorage );
-  gGeoPackageProjectStorage = new QgsGeoPackageProjectStorage;
-  QgsApplication::projectStorageRegistry()->registerProjectStorage( gGeoPackageProjectStorage );  // takes ownership
-}
-
-
-void QgsOgrProviderMetadata::cleanupProvider()
-{
-  QgsApplication::projectStorageRegistry()->unregisterProjectStorage( gGeoPackageProjectStorage );  // destroys the object
-  gGeoPackageProjectStorage = nullptr;
-  QgsOgrConnPool::cleanupInstance();
-  // NOTE: QgsApplication takes care of
-  // calling OGRCleanupAll();
-}
-
-
-
-QgsOgrProviderMetadata::QgsOgrProviderMetadata()
-  : QgsProviderMetadata( TEXT_PROVIDER_KEY, TEXT_PROVIDER_DESCRIPTION )
-{
-
-}
-
-QString QgsOgrProviderMetadata::filters( FilterType type )
-{
-  switch ( type )
-  {
-    case QgsProviderMetadata::FilterType::FilterVector:
-      return QgsOgrProviderUtils::fileVectorFilters();
-
-    case QgsProviderMetadata::FilterType::FilterRaster:
-    case QgsProviderMetadata::FilterType::FilterMesh:
-    case QgsProviderMetadata::FilterType::FilterMeshDataset:
-    case QgsProviderMetadata::FilterType::FilterPointCloud:
-      return QString();
-  }
-  return QString();
-}
-
-bool QgsOgrProviderMetadata::uriIsBlocklisted( const QString &uri ) const
-{
-  const QVariantMap parts = decodeUri( uri );
-  if ( !parts.contains( QStringLiteral( "path" ) ) )
-    return false;
-
-  QFileInfo fi( parts.value( QStringLiteral( "path" ) ).toString() );
-  const QString suffix = fi.completeSuffix();
-
-  // internal details only
-  if ( suffix.compare( QLatin1String( "shp.xml" ), Qt::CaseInsensitive ) == 0 )
-    return true;
-
-  return false;
-}
-
-QMap<QString, QgsAbstractProviderConnection *> QgsOgrProviderMetadata::connections( bool cached )
-{
-  return connectionsProtected<QgsGeoPackageProviderConnection, QgsOgrDbConnection>( cached );
-}
-
-QgsAbstractProviderConnection *QgsOgrProviderMetadata::createConnection( const QString &connName )
-{
-  return new QgsGeoPackageProviderConnection( connName );
-}
-
-QgsAbstractProviderConnection *QgsOgrProviderMetadata::createConnection( const QString &uri, const QVariantMap &configuration )
-{
-  return new QgsGeoPackageProviderConnection( uri, configuration );
-}
-
-void QgsOgrProviderMetadata::deleteConnection( const QString &name )
-{
-  deleteConnectionProtected<QgsGeoPackageProviderConnection>( name );
-}
-
-void QgsOgrProviderMetadata::saveConnection( const QgsAbstractProviderConnection *conn, const QString &name )
-{
-  saveConnectionProtected( conn, name );
-}
-
-QgsProviderMetadata::ProviderCapabilities QgsOgrProviderMetadata::providerCapabilities() const
-{
-  return FileBasedUris | SaveLayerMetadata;
-}
 ///@endcond
